+++ python/damask/_geomgrid.py
@@ -216,7 +216,7 @@ class GeomGrid:
 
         """
         v = VTK.load(fname if str(fname).endswith('.vti') else str(fname)+'.vti')
-        cells = np.array(v.vtk_data.GetDimensions())-1
+        cells = np.array(v.vtk_data.GetDimensions())-1  # type: ignore
         bbox  = np.array(v.vtk_data.GetBounds()).reshape(3,2).T
         ic = {l:v.get(l).reshape(cells,order='F') for l in set(v.labels['Cell Data']) - {label}}
 
@@ -386,7 +386,7 @@ class GeomGrid:
 
         """
         v = VTK.load(fname,'ImageData')
-        cells = np.array(v.vtk_data.GetDimensions())-1
+        cells = np.array(v.vtk_data.GetDimensions())-1 # type: ignore
         bbox  = np.array(v.vtk_data.GetBounds()).reshape(3,2).T
 
         return GeomGrid(material = v.get('MaterialId').reshape(cells,order='F').astype('int32',casting='unsafe'),
