### 2024-07-16 #6986 @MarDiehl
> **License info**:
>
> 🟡 License found: `Other` ([Check here](https://opensource.org/licenses) for OSI approval)
The license is AGPL-3.0-or-later, GitHub does not detect it properly
