### 2024-07-16 #6986 @MarDiehl
For the software report: Since this paper is only about the python code, only the `python` subfolder should be considered for statistics.
