### 2024-07-18 #6986 @Kevin-Mattheus-Moerman
@MarDiehl can you try to resolve this error? :point_up:

If you make any updates to the .bib file, you can call `@editorialbot check references` here again, to check the references again. You can also call `@editorialbot generate pdf` to update the paper once the check is  passing.
