+++ paper.bib
@@ -133,7 +133,7 @@
   publisher = {Elsevier},
   title     = {{ParaView}: An End-User Tool for Large Data Visualization},
   year      = {2005},
-  pages     = 717-731,
+  pages     = {717-731},
   doi      = {10.1016/B978-012387582-2/50038-1},
 }
 
