### 2024-07-22 7b12602
drop support for deprecated ASCII-based geom format
