### 2024-07-26 39fd72c
bugfix in frame transformation; test docstring examples
