### 2024-07-26 fcd1af1
degrees not needed if zero
