### 2024-07-26 5eccae0
vtk subtypes can be annotated with abstract supertypes
