+++ python/damask/_vtk.py
@@ -4,6 +4,9 @@ from pathlib import Path
 from typing import Optional, Union, Literal, List, Sequence
 
 import numpy as np
+
+from vtkmodules.vtkIOXML import vtkXMLReader, vtkXMLWriter
+
 from vtkmodules.vtkCommonCore import (
     vtkPoints,
     vtkStringArray,
@@ -79,7 +82,7 @@ class VTK:
             vtkPolyData, and vtkRectilinearGrid.
 
         """
-        self.vtk_data = vtk_data
+        self.vtk_data: vtkDataSet = vtk_data
 
 
     def __repr__(self) -> str:
@@ -351,39 +354,34 @@ class VTK:
         if not Path(fname).expanduser().is_file():                                                  # vtk has a strange error handling
             raise FileNotFoundError(f'file "{fname}" not found')
         if (ext := Path(fname).suffix) == '.vtk' or dataset_type is not None:
-            reader = vtkGenericDataObjectReader()
-            reader.SetFileName(str(Path(fname).expanduser()))
+            vtk_reader = vtkGenericDataObjectReader()
+            vtk_reader.SetFileName(str(Path(fname).expanduser()))
             if dataset_type is None:
                 raise TypeError('dataset type for *.vtk file not given')
-            elif dataset_type.lower().endswith(('imagedata','image_data')):
-                reader.Update()
-                vtk_data = reader.GetStructuredPointsOutput()
-            elif dataset_type.lower().endswith(('unstructuredgrid','unstructured_grid')):
-                reader.Update()
-                vtk_data = reader.GetUnstructuredGridOutput()
-            elif dataset_type.lower().endswith(('polydata','poly_data')):
-                reader.Update()
-                vtk_data = reader.GetPolyDataOutput()
-            elif dataset_type.lower().endswith(('rectilineargrid','rectilinear_grid')):
-                reader.Update()
-                vtk_data = reader.GetRectilinearGridOutput()
+            vtk_reader.Update()
+            if dataset_type.lower().endswith(('imagedata', 'image_data')):
+                vtk_data = vtk_reader.GetStructuredPointsOutput()
+            elif dataset_type.lower().endswith(('unstructuredgrid', 'unstructured_grid')):
+                vtk_data = vtk_reader.GetUnstructuredGridOutput()
+            elif dataset_type.lower().endswith(('polydata', 'poly_data')):
+                vtk_data = vtk_reader.GetPolyDataOutput()
+            elif dataset_type.lower().endswith(('rectilineargrid', 'rectilinear_grid')):
+                vtk_data = vtk_reader.GetRectilinearGridOutput()
             else:
                 raise TypeError(f'unknown dataset type "{dataset_type}" for vtk file')
         else:
-            if   ext == '.vti':
-                reader = vtkXMLImageDataReader() # type: ignore
-            elif ext == '.vtu':
-                reader = vtkXMLUnstructuredGridReader() # type: ignore
-            elif ext == '.vtp':
-                reader = vtkXMLPolyDataReader() # type: ignore
-            elif ext == '.vtr':
-                reader = vtkXMLRectilinearGridReader() # type: ignore
-            else:
+            xml_reader: Optional[vtkXMLReader] = (
+                vtkXMLImageDataReader() if ext == '.vti' else
+                vtkXMLUnstructuredGridReader() if ext == '.vtu' else
+                vtkXMLPolyDataReader() if ext == '.vtp' else
+                vtkXMLRectilinearGridReader() if ext == '.vtr' else
+                None
+            )
+            if xml_reader is None:
[…]
