+++ python/tests/test_Orientation.py
@@ -257,6 +257,26 @@ class TestOrientation:
         o_2 = Orientation.from_Euler_angles(family=family,phi=[360,0,0],degrees=True)
         assert np.allclose((o_1.disorientation(o_2)).as_matrix(),np.eye(3))
 
+    @pytest.mark.parametrize('shape',[[None,None,()],
+                                      [[2,3,4],[2,3,4],(2,3,4)],
+                                      [[3,4],[4,5],(3,4,5)],
+                                      [[3,2,4],[2,4,6],(3,2,4,6)],
+                                      [[3,4,4],[4,4,2],(3,4,4,2)],
+                                      [100,100,(100,)]])
+    def test_shape_blending(self,shape):
+        o_1 = Orientation.from_random(shape=shape[0],family='triclinic')
+        o_2 = Orientation.from_random(shape=shape[1],family='triclinic')
+        full = o_1.misorientation(o_2).as_axis_angle(pair=True)[1]
+        composition = o_1*o_2
+        assert full.shape == composition.shape == shape[2]
+
+    def test_disorientation_invalid(self):
+        a,b = np.random.choice(list(crystal_families),2,False)
+        o_1 = Orientation.from_random(family=a)
+        o_2 = Orientation.from_random(family=b)
+        with pytest.raises(NotImplementedError):
+            o_1.disorientation(o_2)
+
     @pytest.mark.parametrize('color',[{'label':'red',  'RGB':[1,0,0],'direction':[0,0,1]},
                                       {'label':'green','RGB':[0,1,0],'direction':[0,1,1]},
                                       {'label':'blue', 'RGB':[0,0,1],'direction':[1,1,1]}])
