### 2024-07-26 468c31c
bugfix: correct blending for misorientation
