### 2024-07-27 26251a0
fast calculation of disorientation angle
code is from Lionel Germain, Université de Lorraine.
Publication is under preparation
