### 2024-07-27 50f6bd3
simplification; explicit variable names
