+++ src/grid/grid_mech_spectral_Galerkin.f90
@@ -3,9 +3,9 @@
 !> @author Martin Diehl, Max-Planck-Institut für Eisenforschung GmbH
 !> @author Philip Eisenlohr, Max-Planck-Institut für Eisenforschung GmbH
 !> @author Yi Hu, Max-Planck-Institut für Eisenforschung GmbH
-!> @brief Grid solver for mechanics: Spectral variational
+!> @brief Grid solver for mechanics: Spectral Galerkin
 !--------------------------------------------------------------------------------------------------
-module grid_mechanical_spectral_variational
+module grid_mechanical_spectral_Galerkin
 #include <petsc/finclude/petscsnes.h>
 #include <petsc/finclude/petscdmda.h>
   use PETScDMDA
@@ -89,11 +89,11 @@ module grid_mechanical_spectral_variational
   integer(kind(STATUS_OK)) :: status
 
   public :: &
-    grid_mechanical_spectral_variational_init, &
-    grid_mechanical_spectral_variational_solution, &
-    grid_mechanical_spectral_variational_forward, &
-    grid_mechanical_spectral_variational_updateCoords, &
-    grid_mechanical_spectral_variational_restartWrite
+    grid_mechanical_spectral_Galerkin_init, &
+    grid_mechanical_spectral_Galerkin_solution, &
+    grid_mechanical_spectral_Galerkin_forward, &
+    grid_mechanical_spectral_Galerkin_updateCoords, &
+    grid_mechanical_spectral_Galerkin_restartWrite
 
   ! Missing interfaces for some PETSc versions
   interface MatCreateShell
@@ -160,7 +160,7 @@ contains
 !--------------------------------------------------------------------------------------------------
 !> @brief Allocate all necessary fields and fill them with data, potentially from restart info.
 !--------------------------------------------------------------------------------------------------
-subroutine grid_mechanical_spectral_variational_init(num_grid)
+subroutine grid_mechanical_spectral_Galerkin_init(num_grid)
 
   type(tDict), pointer, intent(in) :: num_grid
 
@@ -183,7 +183,7 @@ subroutine grid_mechanical_spectral_variational_init(num_grid)
   PC  :: pc
 
 
-  print'(/,1x,a)', '<<<+-  grid_mechanical_spectral_variational init  -+>>>'; flush(IO_STDOUT)
+  print'(/,1x,a)', '<<<+-  grid_mechanical_spectral_Galerkin init  -+>>>'; flush(IO_STDOUT)
 
   print'(/,1x,a)', 'J. Vondřejc et al., Computers & Mathematics with Applications 68(3):156–173, 2014'
   print'(  1x,a)', 'https://doi.org/10.1016/j.camwa.2014.05.014'//IO_EOL
@@ -339,13 +339,13 @@ subroutine grid_mechanical_spectral_variational_init(num_grid)
 
   end if restartRead2
 
-end subroutine grid_mechanical_spectral_variational_init
+end subroutine grid_mechanical_spectral_Galerkin_init
 
 
 !--------------------------------------------------------------------------------------------------
-!> @brief solution for the variational scheme with internal iterations
+!> @brief solution for the Galerkin scheme with internal iterations
 !--------------------------------------------------------------------------------------------------
-function grid_mechanical_spectral_variational_solution(incInfoIn) result(solution)
+function grid_mechanical_spectral_Galerkin_solution(incInfoIn) result(solution)
 
 !--------------------------------------------------------------------------------------------------
 ! input data for solution
@@ -374,15 +374,15 @@ function grid_mechanical_spectral_variational_solution(incInfoIn) result(solutio
   solution%iterationsNeeded = totalIter
   P_aim = merge(P_av,P_aim,params%stress_mask)
 
-end function grid_mechanical_spectral_variational_solution
[…]
