+++ src/mesh/discretization_mesh.f90
@@ -71,9 +71,7 @@ contains
 !--------------------------------------------------------------------------------------------------
 !> @brief Initialize the mesh.
 !--------------------------------------------------------------------------------------------------
-subroutine discretization_mesh_init(restart)
-
-  logical, intent(in) :: restart
+subroutine discretization_mesh_init()
 
   PetscInt :: dimPlex, &
     mesh_Nnodes, &                                                                                  !< total number of nodes in mesh
