+++ src/grid/spectral_utilities.f90
@@ -548,7 +548,7 @@ function utilities_G_Convolution(field,stress_mask,fieldAim) result(G_Field)
     i, j, k, &
     l, m
 
-  complex(pREAL), dimension(3,3) :: F_aim                                                           !< zero freq of field
+  complex(pREAL), dimension(3,3) :: field_zero_freq                                                 !< zero freq of field
 
   ! print'(/,1x,a)', '... doing G convolution ...............................................'
   ! flush(IO_STDOUT)
@@ -558,9 +558,9 @@ function utilities_G_Convolution(field,stress_mask,fieldAim) result(G_Field)
   call fftw_mpi_execute_dft_r2c(planTensorForth,tensorField_real,tensorField_fourier)
 
   if (present(fieldAim)) then
-    F_aim = cmplx(fieldAim/wgt,0.0_pREAL,pREAL)
+    field_zero_freq = cmplx(fieldAim/wgt,0.0_pREAL,pREAL)
   else
-    F_aim = tensorField_fourier(1:3,1:3,1,1,1) ! f_hat(k=0) = f_ave * vol = f_ave / wgt
+    field_zero_freq = tensorField_fourier(1:3,1:3,1,1,1) ! f_hat(k=0) = f_ave * vol = f_ave / wgt
   endif
 
   !$OMP PARALLEL DO PRIVATE(l,m,temp33_cmplx)
@@ -582,7 +582,7 @@ function utilities_G_Convolution(field,stress_mask,fieldAim) result(G_Field)
 
   ! Apply stress bounday conditions
   if (cells3Offset == 0) &
-    tensorField_fourier(1:3,1:3,1,1,1) = merge(cmplx(0.0_pREAL,0.0_pREAL,pREAL),F_aim,stress_mask)
+    tensorField_fourier(1:3,1:3,1,1,1) = merge(cmplx(0.0_pREAL,0.0_pREAL,pREAL),field_zero_freq,stress_mask)
 
 
   call fftw_mpi_execute_dft_c2r(planTensorBack,tensorField_fourier,tensorField_real)
