### 2024-07-31 322a93e
need to reduce precision for certain cases
