+++ .gitlab/issue_templates/new_release.md
@@ -1,0 +1,20 @@
+The following tasks are required for a new release.
+Detailed instructions for each step are documented in the wiki
+
+1. [ ] Compile changelog
+1. [ ] Update repository
+1. [ ] Create and upload tarball
+1. [ ] Create GitLab release
+1. [ ] Update examples repository
+1. [ ] Update website
+1. Packages
+  - [ ] Conda package
+  - [ ] Spack package
+  - [ ] PyPi package (wheel)
+  - [ ] Arch Linux (PKGBUILD)
+  - [ ] Ubuntu (Launchpad)
+  - Open build service
+    - [ ] Debian
+    - [ ] Fedora/Suse
+1. [ ] Docker
+1. [ ] Send info mail to mailing list
