### 2024-08-07 87615a4
Update new_release.md
