+++ .gitlab/issue_templates/new_release.md
@@ -8,13 +8,13 @@ Detailed instructions for each step are documented in the wiki
 1. [ ] Update examples repository
 1. [ ] Update website
 1. Packages
-  - [ ] Conda package
-  - [ ] Spack package
-  - [ ] PyPi package (wheel)
-  - [ ] Arch Linux (PKGBUILD)
-  - [ ] Ubuntu (Launchpad)
-  - Open build service
-    - [ ] Debian
-    - [ ] Fedora/Suse
+   - [ ] Conda package
+   - [ ] Spack package
+   - [ ] PyPi package (wheel)
+   - [ ] Arch Linux (PKGBUILD)
+   - [ ] Ubuntu (Launchpad)
+   - Open build service
+     - [ ] Debian
+     - [ ] Fedora/Suse
 1. [ ] Docker
 1. [ ] Send info mail to mailing list
