### 2024-08-07 6248521
Update new_release.md
