### 2024-08-12 8f7eb53
less intrusive formatting of quaternion scalar and vector parts
