+++ python/damask/_rotation.py
@@ -97,7 +97,7 @@ class Rotation:
         """
         s = self.quaternion.__repr__()[6:-1].replace('\n'+' '*6,'\n')
         return re.sub(r'\[([^\[\]]*),\n? ?([^\[\]]*),\n? ?([^\[\]]*),\n? ?([^\[\]]*)\]',
-                      r'real=\1, imag=<\2,\3,\4>',s)
+                      r'\1,    \2,\3,\4',s)
 
 
     def __copy__(self: MyType,
