+++ .gitlab-ci.yml
@@ -67,7 +67,7 @@ setuptools:
   stage: python
   tags:
     - matesting2-container-runner
-  image: git.damask.mpie.de:5050/damask/damask:python310_2407
+  image: git.damask-multiphysics.org:5050/damask/damask:python310_2407
   before_script: []
   script:
     - sed -i 's/-[[:digit:]]*-.*//' VERSION
@@ -78,7 +78,7 @@ pytest:
   stage: python
   tags:
     - matesting2-container-runner
-  image: git.damask.mpie.de:5050/damask/damask:python310_2407
+  image: git.damask-multiphysics.org:5050/damask/damask:python310_2407
   before_script: []
   script:
     - PYTHONPATH=$(pwd)/python:$PYTHONPATH
@@ -90,7 +90,7 @@ mypy:
   stage: python
   tags:
     - matesting2-container-runner
-  image: git.damask.mpie.de:5050/damask/damask:python310_2407
+  image: git.damask-multiphysics.org:5050/damask/damask:python310_2407
   before_script: []
   script:
     - cd python
@@ -256,7 +256,7 @@ grid_performance:
     - make -j2 all install
     - export PATH=${PWD}/bin:${PATH}
     - cd $(mktemp -d)
-    - git clone -q git@git.damask.mpie.de:damask/statistics.git .
+    - git clone -q git@git-multiphysics.org:damask/statistics.git .
     - ./measure_performance.py --input_dir ${CI_PROJECT_DIR}/examples/grid --tag ${CI_COMMIT_SHA}
     - >
         if [ ${CI_COMMIT_BRANCH} == development ]; then
@@ -282,7 +282,7 @@ update_revision:
     - echo Job start:" $(date)"
   script:
     - cd $(mktemp -d)
-    - git clone -q git@git.damask.mpie.de:damask/DAMASK.git .
+    - git clone -q git@git.damask-multiphysics.org:damask/DAMASK.git .
     - git pull
     - export VERSION=$(git describe ${CI_COMMIT_SHA})
     - echo ${VERSION:1} > VERSION
