### 2024-08-20 #6986 @MarDiehl
- sergiolucarini
- hakonanes
- rquey
- MHenrich1990
- s5409693 (recent publication on different topic)
- mkuehbach (former colleague, joint publication in 2020)
