### 2024-08-26 08a7a21
helps to understand the table
see, e.g.
https://en.wikipedia.org/wiki/Crystal_system#Crystal_classes
