+++ python/damask/_crystal.py
@@ -719,7 +719,7 @@ class Crystal():
                               [-0.5*np.sqrt(2), 0.0,           -0.5*np.sqrt(2), 0.0            ],
                               [-0.5*np.sqrt(2), 0.5*np.sqrt(2), 0.0,            0.0            ],
                               [-0.5*np.sqrt(2),-0.5*np.sqrt(2), 0.0,            0.0            ],
-                            ],
+                            ], # 432
             'hexagonal':    [
                               [ 1.0,            0.0,            0.0,            0.0            ],
                               [-0.5*np.sqrt(3), 0.0,            0.0,           -0.5            ],
@@ -733,7 +733,7 @@ class Crystal():
                               [ 0.0,            0.0,            1.0,            0.0            ],
                               [ 0.0,           -0.5,           -0.5*np.sqrt(3), 0.0            ],
                               [ 0.0,            0.5*np.sqrt(3), 0.5,            0.0            ],
-                            ],
+                            ], # 622
             'tetragonal':   [
                               [ 1.0,            0.0,            0.0,            0.0            ],
                               [ 0.0,            1.0,            0.0,            0.0            ],
@@ -743,20 +743,20 @@ class Crystal():
                               [ 0.0,           -0.5*np.sqrt(2), 0.5*np.sqrt(2), 0.0            ],
                               [ 0.5*np.sqrt(2), 0.0,            0.0,            0.5*np.sqrt(2) ],
                               [-0.5*np.sqrt(2), 0.0,            0.0,            0.5*np.sqrt(2) ],
-                            ],
+                            ], # 422
             'orthorhombic': [
                               [ 1.0,0.0,0.0,0.0 ],
                               [ 0.0,1.0,0.0,0.0 ],
                               [ 0.0,0.0,1.0,0.0 ],
                               [ 0.0,0.0,0.0,1.0 ],
-                            ],
+                            ], # 222
             'monoclinic':   [
                               [ 1.0,0.0,0.0,0.0 ],
                               [ 0.0,0.0,1.0,0.0 ],
-                            ],
+                            ], # 2
             'triclinic':    [
                               [ 1.0,0.0,0.0,0.0 ],
-                            ]}
+                            ]} # 1
         return Rotation.from_quaternion(_symmetry_operations[self.family],accept_homomorph=True)
 
 
