### 2024-08-26 18eeecb
add missing restore, read-only access suffices
