+++ python/damask/_rotation.py
@@ -96,7 +96,6 @@ class Rotation:
 
         """
         s = self.quaternion.__repr__()[6:-1].replace('\n'+' '*6,'\n')
-        print(f'debug:\n{self.quaternion.__repr__()}\n{s}')
         return re.sub(r'\[([^\[\]]*),(\n )? ?([^\[\]]*),(\n )? ?([^\[\]]*),(\n )? ?([^\[\]]*)\]',
                       r'\1,    \3,\5,\7',s)
 
