### 2024-08-29 e8b00bc
improved readability
RestoreTabulation is probably not needed, but does not hurt either
