### 2024-08-30 bf86bcb
robust __str__ and __repr__
