### 2024-09-02 #7164 @editorialbot
[REVIEW]: A Python Library for Pre- and Post-Processing of DAMASK Simulations
**Submitting author:** @MarDiehl (<a href="http://orcid.org/0000-0002-3738-7363">Martin Diehl</a>)
**Repository:** https://github.com/damask-multiphysics/DAMASK
**Branch with paper.md** (empty if default branch): JOSS-paper
**Version:** v3.0.0-beta2
