### 2024-09-02 #7164 @zhubonan
Hi @hakonanes and  @mkuehbach, thanks for agreeing to review this submission 👏   - please create your checklist by typing:

To make comments and suggestions (as in normal peer-review journals) to the author, you can directly raise issues in their code's GitHub repo and mention this page so we know that it is linked to JOSS.

I would like to aim for having the review completed in two weeks. Please do let me know if you need more time.
