### 2024-09-02 #7164 @zhubonan
@MarDiehl it seems that there are a few missing DOIs in the manuscript, could you please fix those?
