### 2024-09-02 #7164 @MarDiehl
> @MarDiehl it seems that there are a few missing DOIs in the manuscript, could you please fix those?

These are false positives from entries in the bib file that are not cited. From all publications cited in the submitted version of the papers, only _Collette, A. (2013). Python and HDF5. O’Reilly. ISBN: 9781449367831_ does not have a DOI because I don't think it has one. ISBN should anyways do for a book.
