### 2024-09-02 #7164 @MarDiehl
> **License info**:
>
> 🟡 License found: `Other` ([Check here](https://opensource.org/licenses) for OSI approval)

License is `AGPL-3.0-or-later` but GitHub does not detect it properly (GitLab does).
