### 2024-09-04 2216bb5
backport for NumPy's np.broadcast_shapes
not available in versions < 1.20.0, Ubuntu 20.04 has NumPy 1.17.x
