### 2024-09-05 28dd26e
would lead to infinite strain rate
