+++ src/grid/DAMASK_grid.f90
@@ -97,15 +97,15 @@ program DAMASK_grid
   type(tLoadCase), allocatable, dimension(:) :: loadCases                                           !< array of all load cases
   type(tSolutionState), allocatable, dimension(:) :: solres
   procedure(grid_mechanical_spectral_basic_init), pointer :: &
-    mechanical_init
+    grid_mechanical_init
   procedure(grid_mechanical_spectral_basic_forward), pointer :: &
-    mechanical_forward
+    grid_mechanical_forward
   procedure(grid_mechanical_spectral_basic_solution), pointer :: &
-    mechanical_solution
+    grid_mechanical_solution
   procedure(grid_mechanical_spectral_basic_updateCoords), pointer :: &
-    mechanical_updateCoords
+    grid_mechanical_updateCoords
   procedure(grid_mechanical_spectral_basic_restartWrite), pointer :: &
-    mechanical_restartWrite
+    grid_mechanical_restartWrite
 
   external :: &
     quit
@@ -160,35 +160,35 @@ program DAMASK_grid
   nActiveFields = 1
   select case (solver%get_asStr('mechanical'))
     case ('spectral_basic')
-      mechanical_init         => grid_mechanical_spectral_basic_init
-      mechanical_forward      => grid_mechanical_spectral_basic_forward
-      mechanical_solution     => grid_mechanical_spectral_basic_solution
-      mechanical_updateCoords => grid_mechanical_spectral_basic_updateCoords
-      mechanical_restartWrite => grid_mechanical_spectral_basic_restartWrite
+      grid_mechanical_init         => grid_mechanical_spectral_basic_init
+      grid_mechanical_forward      => grid_mechanical_spectral_basic_forward
+      grid_mechanical_solution     => grid_mechanical_spectral_basic_solution
+      grid_mechanical_updateCoords => grid_mechanical_spectral_basic_updateCoords
+      grid_mechanical_restartWrite => grid_mechanical_spectral_basic_restartWrite
       active_Gamma = .true.
 
     case ('spectral_polarization')
-      mechanical_init         => grid_mechanical_spectral_polarization_init
-      mechanical_forward      => grid_mechanical_spectral_polarization_forward
-      mechanical_solution     => grid_mechanical_spectral_polarization_solution
-      mechanical_updateCoords => grid_mechanical_spectral_polarization_updateCoords
-      mechanical_restartWrite => grid_mechanical_spectral_polarization_restartWrite
+      grid_mechanical_init         => grid_mechanical_spectral_polarization_init
+      grid_mechanical_forward      => grid_mechanical_spectral_polarization_forward
+      grid_mechanical_solution     => grid_mechanical_spectral_polarization_solution
+      grid_mechanical_updateCoords => grid_mechanical_spectral_polarization_updateCoords
+      grid_mechanical_restartWrite => grid_mechanical_spectral_polarization_restartWrite
       active_Gamma = .true.
 
     case ('spectral_Galerkin')
-      mechanical_init         => grid_mechanical_spectral_Galerkin_init
-      mechanical_forward      => grid_mechanical_spectral_Galerkin_forward
-      mechanical_solution     => grid_mechanical_spectral_Galerkin_solution
-      mechanical_updateCoords => grid_mechanical_spectral_Galerkin_updateCoords
-      mechanical_restartWrite => grid_mechanical_spectral_Galerkin_restartWrite
+      grid_mechanical_init         => grid_mechanical_spectral_Galerkin_init
+      grid_mechanical_forward      => grid_mechanical_spectral_Galerkin_forward
+      grid_mechanical_solution     => grid_mechanical_spectral_Galerkin_solution
+      grid_mechanical_updateCoords => grid_mechanical_spectral_Galerkin_updateCoords
[…]
