### 2024-09-14 4d4da9a
consistent definition of twinning shear signs
note: the sign of the characteristic shear indicates whether a hP twin
is a tension twin (uniaxial stress along c causes extension) or a
compression twin (uniaxial stress along -c causes shortening). Depending
on that, the sign of the plane is flipped. I'm not 100% sure whether
this is correct.
In the constitutive laws, characteristic shear needs to be positive.
