+++ src/crystal.f90
@@ -261,7 +261,7 @@ module crystal
     ! -----------------------------------
 
     ! <-10.1>{10.2} systems, shear = (3-(c/a)^2)/(sqrt(3) c/a)
-    ! tension in Co, Mg, Zr, Ti, and Be; compression in Cd and Zn
+    ! tension for c/a < sqrt(3): Mg, Ti, Co, Zr, and Be; compression for c/a > sqrt(3): Zn and Cd
       -1,  0,  1,  1,     1,  0, -1,  2, & !
        0, -1,  1,  1,     0,  1, -1,  2, &
        1, -1,  0,  1,    -1,  1,  0,  2, &
@@ -269,7 +269,7 @@ module crystal
        0,  1, -1,  1,     0, -1,  1,  2, &
       -1,  1,  0,  1,     1, -1,  0,  2, &
     ! <11.6>{-1-1.1} systems, shear = 1/(c/a)
-    ! tension in Co, Re, and Zr
+    ! tension: Ti, Re, Zr, Co, and graphite
       -1, -1,  2,  6,     1,  1, -2,  1, &
        1, -2,  1,  6,    -1,  2, -1,  1, &
        2, -1, -1,  6,    -2,  1,  1,  1, &
@@ -277,22 +277,23 @@ module crystal
       -1,  2, -1,  6,     1, -2,  1,  1, &
       -2,  1,  1,  6,     2, -1, -1,  1, &
     ! <10.-2>{10.1} systems, shear = (4(c/a)^2-9)/(4 sqrt(3) c/a)
-    ! compression in Mg
-       1,  0, -1, -2,     1,  0, -1,  1, &
-       0,  1, -1, -2,     0,  1, -1,  1, &
-      -1,  1,  0, -2,    -1,  1,  0,  1, &
-      -1,  0,  1, -2,    -1,  0,  1,  1, &
-       0, -1,  1, -2,     0, -1,  1,  1, &
-       1, -1,  0, -2,     1, -1,  0,  1, &
+    ! compression for c/a > 1.5: Mg and Re
+       1,  0, -1, -2,    -1, -0,  1, -1, &
+       0,  1, -1, -2,    -0, -1,  1, -1, &
+      -1,  1,  0, -2,     1, -1, -0, -1, &
+      -1,  0,  1, -2,     1, -0, -1, -1, &
+       0, -1,  1, -2,    -0,  1, -1, -1, &
+       1, -1,  0, -2,    -1,  1, -0, -1, &
     ! <11.-3>{11.2} systems, shear = 2((c/a)^2-2)/(3 c/a)
-    ! compression in Ti and Zr
-       1,  1, -2, -3,     1,  1, -2,  2, &
-      -1,  2, -1, -3,    -1,  2, -1,  2, &
-      -2,  1,  1, -3,    -2,  1,  1,  2, &
-      -1, -1,  2, -3,    -1, -1,  2,  2, &
-       1, -2,  1, -3,     1, -2,  1,  2, &
-       2, -1, -1, -3,     2, -1, -1,  2  &
+    ! compression for c/a > sqrt(2): Ti and Zr
+       1,  1, -2, -3,    -1, -1,  2, -2, &
+      -1,  2, -1, -3,     1, -2,  1, -2, &
+      -2,  1,  1, -3,     2, -1, -1, -2, &
+      -1, -1,  2, -3,     1,  1, -2, -2, &
+       1, -2,  1, -3,    -1,  2, -1, -2, &
+       2, -1, -1, -3,    -2,  1,  1, -2  &
       ],pREAL),shape(HP_SYSTEMTWIN))                                                                !< hP twin systems, sorted by P. Eisenlohr CCW around <c> starting next to a_1 axis
[…]
