### 2024-09-16 79d441e
always include all possible output items in examples
also ensure that the output is valid (in PRIVATE tests)
