+++ examples/config/phase/mechanical/plastic/phenopowerlaw_bcc-martensite.yaml
@@ -9,6 +9,8 @@ references:
     Metallurgical and Materials Transactions B 1:1121–1143, 1970,
     https://doi.org/10.1007/BF02900224
 
+output: [xi_sl, gamma_sl]
+
 N_sl: [12, 12]
 
 dot_gamma_0_sl: [0.001, 0.001]
