+++ examples/config/homogenization/mechanical/pass_direct.yaml
@@ -1,3 +1,4 @@
 # For single point calculations, requires N_constituents = 1
 type: pass
-output: ['F', 'P']
+
+output: [F, P]
