+++ examples/config/phase/thermal/X2CrNi18-10.yaml
@@ -5,6 +5,8 @@ references:
     https://doi.org/10.2172/4152287,
     Eq. 5 and Eq. 28 (T_min=300K, T_max=1600K)
 
+output: [T]
+
 K_11: 12.86
 K_11,T: 1.618e-2
 
