+++ examples/config/phase/mechanical/plastic/kinehardening_X2CrNiMo18-15-4.yaml
@@ -5,7 +5,7 @@ references:
     International Journal of Fatigue 36(1):181-193, 2012,
     https://doi.org/10.1016/j.ijfatigue.2011.07.008
 
-output: [xi, chi, chi_flip, gamma_flip, gamma, sgn(gamma)]
+output: [xi, chi, sgn(gamma), chi_flip, gamma_flip, gamma]
 
 N_sl: [12]
 
