+++ src/crystal.f90
@@ -276,7 +276,7 @@ module crystal
        1,  1, -2,  6,    -1, -1,  2,  1, &
       -1,  2, -1,  6,     1, -2,  1,  1, &
       -2,  1,  1,  6,     2, -1, -1,  1, &
-    ! <10.-2>{10.1} systems, shear = (4(c/a)^2-9)/(4 sqrt(3) c/a)
+    ! <10.-2>{10.1} systems, shear = (9-4(c/a)^2)/(4 sqrt(3) c/a)
     ! compression for c/a > 1.5: Mg and Re
        1,  0, -1, -2,    -1, -0,  1, -1, &
        0,  1, -1, -2,    -0, -1,  1, -1, &
@@ -284,7 +284,7 @@ module crystal
       -1,  0,  1, -2,     1, -0, -1, -1, &
        0, -1,  1, -2,    -0,  1, -1, -1, &
        1, -1,  0, -2,    -1,  1, -0, -1, &
-    ! <11.-3>{11.2} systems, shear = 2((c/a)^2-2)/(3 c/a)
+    ! <11.-3>{11.2} systems, shear = 2(2-(c/a)^2)/(3 c/a)
     ! compression for c/a > sqrt(2): Ti and Zr
        1,  1, -2, -3,    -1, -1,  2, -2, &
       -1,  2, -1, -3,     1, -2,  1, -2, &
