+++ python/damask/_crystal.py
@@ -1057,28 +1057,28 @@ class Crystal():
                            [-1,+2,-1,+3, +1,-2,+1,+2],
                            [-2,+1,+1,+3, +2,-1,-1,+2]])],
                 'twin': [np.array([
-                           [-1, 0, 1, 1,  1, 0,-1, 2],   # shear = (3-(c/a)^2)/(sqrt(3) c/a) <-10.1>{10.2}
+                           [-1, 0, 1, 1,  1, 0,-1, 2],   # <-10.1>{10.2}; shear = (3-(c/a)^2)/(sqrt(3) c/a)
                            [ 0,-1, 1, 1,  0, 1,-1, 2],
                            [ 1,-1, 0, 1, -1, 1, 0, 2],
                            [ 1, 0,-1, 1, -1, 0, 1, 2],
                            [ 0, 1,-1, 1,  0,-1, 1, 2],
                            [-1, 1, 0, 1,  1,-1, 0, 2]]),
                          np.array([
-                           [-1,-1, 2, 6,  1, 1,-2, 1],   # shear = 1/(c/a) <11.6>{-1-1.1}
+                           [-1,-1, 2, 6,  1, 1,-2, 1],   # <11.6>{-1-1.1}; shear = 1/(c/a)
                            [ 1,-2, 1, 6, -1, 2,-1, 1],
                            [ 2,-1,-1, 6, -2, 1, 1, 1],
                            [ 1, 1,-2, 6, -1,-1, 2, 1],
                            [-1, 2,-1, 6,  1,-2, 1, 1],
                            [-2, 1, 1, 6,  2,-1,-1, 1]]),
                          np.array([
-                           [ 1, 0,-1,-2, -1,-0, 1,-1],   # shear = (4(c/a)^2-9)/(4 sqrt(3) c/a)  <10.-2>{10.1}
+                           [ 1, 0,-1,-2, -1,-0, 1,-1],   # <10.-2>{10.1}; shear = (9-4(c/a)^2)/(4 sqrt(3) c/a)
                            [ 0, 1,-1,-2, -0,-1, 1,-1],
                            [-1, 1, 0,-2,  1,-1,-0,-1],
                            [-1, 0, 1,-2,  1,-0,-1,-1],
                            [ 0,-1, 1,-2, -0, 1,-1,-1],
                            [ 1,-1, 0,-2, -1, 1,-0,-1]]),
                          np.array([
-                           [ 1, 1,-2,-3, -1,-1, 2,-2],   # shear = 2((c/a)^2-2)/(3 c/a)  <11.-3>{11.2}
+                           [ 1, 1,-2,-3, -1,-1, 2,-2],   # <11.-3>{11.2}; shear = 2(2-(c/a)^2)/(3 c/a)
                            [-1, 2,-1,-3,  1,-2, 1,-2],
                            [-2, 1, 1,-3,  2,-1,-1,-2],
                            [-1,-1, 2,-3,  1, 1,-2,-2],
