### 2024-09-16 a7eb1c1
more extensive documentation
