### 2024-09-17 f48d9e8
polishing
restored hints on metals in which specific twin families are active.
This is not only based on c/a
