+++ python/damask/solver/_marc.py
@@ -4,7 +4,7 @@ import re
 from pathlib import Path
 from typing import Literal
 
-_marc_version = '2023.4'
+_marc_version = '2024.1'
 _marc_root = '/opt/msc'
 _damask_root = str(Path(__file__).parents[3])
 
