### 2024-09-21 3987053
using specific colors for mesh and grid solver
