+++ src/CLI.f90
@@ -78,22 +78,24 @@ subroutine CLI_init()
  ! http://patorjk.com/software/taag/#p=display&f=Lean&t=DAMASK%203
 #ifdef DEBUG
   print'(a)', IO_color([255,0,0])
-  print'(1x,a,/)', 'debug version - debug version - debug version - debug version - debug version'
-#else
-  print'(a)', IO_color([67,128,208])
+  print'(1x,a)', 'debug version - debug version - debug version - debug version - debug version'
 #endif
+  print'(a)', IO_color([67,128,208])
   print'(1x,a)', '    _/_/_/      _/_/    _/      _/    _/_/      _/_/_/  _/    _/    _/_/_/'
   print'(1x,a)', '   _/    _/  _/    _/  _/_/  _/_/  _/    _/  _/        _/  _/            _/'
   print'(1x,a)', '  _/    _/  _/_/_/_/  _/  _/  _/  _/_/_/_/    _/_/    _/_/          _/_/'
   print'(1x,a)', ' _/    _/  _/    _/  _/      _/  _/    _/        _/  _/  _/            _/'
   print'(1x,a)', '_/_/_/    _/    _/  _/      _/  _/    _/  _/_/_/    _/    _/    _/_/_/'
 #if   defined(GRID)
+  print'(a)', IO_color([123,207,68])
   print'(1x,a)', 'Grid solver'
 #elif defined(MESH)
+  print'(a)', IO_color([230,150,68])
   print'(1x,a)', 'Mesh solver'
 #endif
 #ifdef DEBUG
-  print'(/,1x,a)', 'debug version - debug version - debug version - debug version - debug version'
+  print'(a)', IO_color([255,0,0])
+  print'(1x,a)', 'debug version - debug version - debug version - debug version - debug version'
 #endif
   print'(a)', IO_color()
 
