### 2024-09-22 751f2c7
grid solver adjustments for PETSc 3.22.x
and a few smaller adjustments
