### 2024-09-23 57bac69
Replace all uses of the module FEM_quadrature with calls to PETSc functions; in particular:
- PetscDTSimplexQuadrature: creates a quadrature rule.
- PetscQuadratureGetData: retrieve quadrature rule values (integration point coordinates and weights).
    
Needs PETScDT and works only for PETSc > 3.18.
