+++ src/mesh/FEM_utilities.f90
@@ -17,7 +17,9 @@ module FEM_utilities
   use parallelization
   use discretization_mesh
   use homogenization
+#if (PETSC_VERSION_MAJOR==3 && PETSC_VERSION_MINOR<18)
   use FEM_quadrature
+#endif
   use constants
 
 #if (PETSC_VERSION_MAJOR==3 && PETSC_VERSION_MINOR>14) && !defined(PETSC_HAVE_MPI_F90MODULE_VISIBILITY)
@@ -89,7 +91,11 @@ subroutine FEM_utilities_init(num_mesh)
   p_s = num_mesh%get_asInt('p_s',defaultVal = 2)
   p_i = num_mesh%get_asInt('p_i',defaultVal = p_s)
 
+#if (PETSC_VERSION_MAJOR==3 && PETSC_VERSION_MINOR<18)
   if (p_s < 1 .or. p_s > size(FEM_nQuadrature,2)) &
+#else
+  if (p_s < 1) &
+#endif
     call IO_error(821,ext_msg='shape function order (p_s) out of bounds')
   if (p_i < max(1,p_s-1) .or. p_i > p_s) &
     call IO_error(821,ext_msg='integration order (p_i) out of bounds')
