### 2024-09-25 c388cab
from_parallel consistent with from_basis
