### 2024-10-06 #7164 @hakonanes
I noticed a couple of typos (by draft line-number):
* 43: _datas structures_ -> _data structures_
* 63: The sentence _This flexible file format is called DADF5, short for "DAMASK HDF5" ensures FAIR data storage (Wilkinson et al., 2016)_ should have better punctuation. Perhaps _This flexible file format is called DADF5, short for "DAMASK HDF5", and ensures FAIR data storage (Wilkinson et al., 2016)._
