### 2024-10-08 e689d6b
needed to compile with GCC in debug mode
