### 2024-10-08 bd2e4b0
improved handling of optional arguments
1) can be (arbitrarly shaped) arrays now due to "elemental" attribute
2) complex type now supported
