### 2024-10-08 10af4b8
mesh solver adjustments for PETSc 3.22.x
