+++ src/CLI.f90
@@ -6,7 +6,7 @@
 !> @brief Parse command line interface for PETSc-based solvers
 !--------------------------------------------------------------------------------------------------
 #define PETSC_MINOR_MIN 12
-#define PETSC_MINOR_MAX 21
+#define PETSC_MINOR_MAX 22
 
 module CLI
   use, intrinsic :: ISO_fortran_env
