### 2024-10-12 damask-multiphysics/DAMASK#36 @MarDiehl
thanks.
We plan to drop support for Python 3.8 in the DAMASK 3.1.0 release. We still support it in 3.0.0 due to compatibility with Ubuntu 20.04.
