### 2024-10-14 #7164 @hakonanes
I've completed my review, and find the paper fit for publication.

I think users and those about to start using DAMASK will find the Python pre- and post-processing tools very useful (and already are, I'm sure). They are well tested and documented. (Also providing functionality I hope we can implement in orix at some point as well!)

I've raised two questions related to installation in the DAMASK repo. One is already addressed, the other need not be before publication.

*  https://github.com/damask-multiphysics/DAMASK/issues/36
* https://github.com/damask-multiphysics/DAMASK/discussions/37
