### 2024-10-19 e8ad7b3
test with newest PETSc on GitHub actions
