### 2024-10-22 24d767c
bump minimum Python version to 3.9
