+++ python/setup.cfg
@@ -2,8 +2,8 @@
 name = damask
 version = file: damask/VERSION
 author = The DAMASK team
-author_email = damask@mpie.de
-url = https://damask.mpie.de
+author_email = support@damask-multiphyiscs.org
+url = https://damask-multiphysics.org
 description = DAMASK processing tools
 long_description = Pre- and post-processing tools for DAMASK
 license = AGPL3
@@ -17,7 +17,7 @@ classifiers =
 [options]
 packages = find:
 include_package_data = true
-python_requires = >= 3.8
+python_requires = >= 3.9
 install_requires =
     pandas>=0.24                                                                                    # requires numpy
     numpy>=1.17                                                                                     # needed for default_rng
