### 2024-10-22 ab26b43
proper iteration counting
