### 2024-10-23 066c494
avoid division by zero
happens for all-zero tensor (e.g. during initial call to dislotwin/shear
banding).
