### 2024-10-24 e820e34
consider ordering of eigenvectors
