+++ src/math.f90
@@ -1472,15 +1472,18 @@ subroutine math_selfTest()
   call math_eigh33(v3_1,t33_2,t33)
   if (any(dNeq0(v3_1)))         error stop 'math_eigh33/zero eigenvalues (values)'
   if (any(dNeq(t33_2,math_I3))) error stop 'math_eigh33/zero eigenvalues (vectors)'
+
   t33 = math_I3
   call random_number(r)
-  d = int(r*2.0_pREAL) + 1
+  d = nint(r*2.0_pREAL) + 1
   t33(d,d) = 5.0_pREAL + r*10.0_pREAL
   t33(mod(d,3)+1,mod(d,3)+1) = 20.0_pREAL + r*10.0_pREAL
   call math_eigh33(v3_1,t33_2,t33)
   if (any(dNeq(v3_1,[1.0_pREAL,t33(d,d),t33(mod(d,3)+1,mod(d,3)+1)]))) &
                                 error stop 'math_eigh33/non-zero eigenvalues (values)'
-  if (any(dNeq(t33_2,math_I3))) error stop 'math_eigh33/non-zero eigenvalues (vectors)'
+  if (any(dNeq(math_I3(1:3,mod(d+1,3)+1),t33_2(1:3,1)))) error stop 'math_eigh33/min eigenvector'
+  if (any(dNeq(math_I3(1:3,d           ),t33_2(1:3,2)))) error stop 'math_eigh33/mid eigenvector'
+  if (any(dNeq(math_I3(1:3,mod(d,3)+1  ),t33_2(1:3,3)))) error stop 'math_eigh33/max eigenvector'
 
 end subroutine math_selfTest
 
