+++ examples/config/phase/mechanical/elastic/Hooke_X6CrNiMo17-12-2.yaml
@@ -6,13 +6,13 @@ references:
     https://doi.org/10.1007/s11661-006-0129-4,
     fit to Fig. 6 (solid lines, T_min=0°C, T_max=600°C)
 
-C_11: 118.6e+9
-C_11,T: -5.491e+7
+C_11: 229.0e+9
+C_11,T: -1.058e+8
 
 C_12: 154.2e+9
 C_12,T: -6.059e+7
 
-C_44: 229.0e+9
-C_44,T: -1.058e+8
+C_44: 118.6e+9
+C_44,T: -5.491e+7
 
 T_ref: 293.15
