### 2024-10-29 7d83b9b
important difference
https://doi.org/10.1080/02670836.2016.1231746
http://pajarito.materials.cmu.edu/lectures/L23-Metal_Textures-fcc_bcc-part1.pdf
