### 2024-10-31 be306a7
static data
