+++ python/damask/_crystal.py
@@ -7,6 +7,243 @@ from . import util
 from . import Rotation
 
 
+_kinematics: Dict[Optional[BravaisLattice], Dict[CrystalKinematics, List[np.ndarray]]] = {
+    'cF': {
+        'slip': [np.array([
+                   [ 0,+1,-1, +1,+1,+1],
+                   [-1, 0,+1, +1,+1,+1],
+                   [+1,-1, 0, +1,+1,+1],
+                   [ 0,-1,-1, -1,-1,+1],
+                   [+1, 0,+1, -1,-1,+1],
+                   [-1,+1, 0, -1,-1,+1],
+                   [ 0,-1,+1, +1,-1,-1],
+                   [-1, 0,-1, +1,-1,-1],
+                   [+1,+1, 0, +1,-1,-1],
+                   [ 0,+1,+1, -1,+1,-1],
+                   [+1, 0,-1, -1,+1,-1],
+                   [-1,-1, 0, -1,+1,-1]]),
+                 np.array([
+                   [+1,+1, 0, +1,-1, 0],
+                   [+1,-1, 0, +1,+1, 0],
+                   [+1, 0,+1, +1, 0,-1],
+                   [+1, 0,-1, +1, 0,+1],
+                   [ 0,+1,+1,  0,+1,-1],
+                   [ 0,+1,-1,  0,+1,+1]])],
+        'twin': [np.array([
+                   [-2, 1, 1,  1, 1, 1],
+                   [ 1,-2, 1,  1, 1, 1],
+                   [ 1, 1,-2,  1, 1, 1],
+                   [ 2,-1, 1, -1,-1, 1],
+                   [-1, 2, 1, -1,-1, 1],
+                   [-1,-1,-2, -1,-1, 1],
+                   [-2,-1,-1,  1,-1,-1],
+                   [ 1, 2,-1,  1,-1,-1],
+                   [ 1,-1, 2,  1,-1,-1],
+                   [ 2, 1,-1, -1, 1,-1],
+                   [-1,-2,-1, -1, 1,-1],
+                   [-1, 1, 2, -1, 1,-1]])]
+    },
+    'cI': {
+        'slip': [np.array([
+                   [+1,-1,+1, +0,+1,+1],
+                   [+1,-1,+1, +1,+0,-1],
+                   [+1,-1,+1, -1,-1,+0],
+                   [-1,-1,+1, +0,-1,-1],
+                   [-1,-1,+1, +1,+0,+1],
+                   [-1,-1,+1, -1,+1,+0],
+                   [+1,+1,+1, +0,+1,-1],
+                   [+1,+1,+1, -1,+0,+1],
+                   [+1,+1,+1, +1,-1,+0],
+                   [-1,+1,+1, +0,-1,+1],
+                   [-1,+1,+1, -1,+0,-1],
+                   [-1,+1,+1, +1,+1,+0]]),
+                 np.array([
+                   [+1,-1,+1, +2,+1,-1],
+                   [+1,-1,+1, -1,+1,+2],
+                   [+1,-1,+1, +1,+2,+1],
+                   [-1,-1,+1, +2,-1,+1],
+                   [-1,-1,+1, +1,+1,+2],
+                   [-1,-1,+1, -1,+2,+1],
+                   [+1,+1,+1, +1,+1,-2],
+                   [+1,+1,+1, +1,-2,+1],
+                   [+1,+1,+1, -2,+1,+1],
+                   [-1,+1,+1, +1,-1,+2],
+                   [-1,+1,+1, +1,+2,-1],
+                   [-1,+1,+1, +2,+1,+1]]),
+                 np.array([
+                   [+1,-1,+1, -1,+2,+3],
[…]
