### 2024-11-01 073d0f0
array creation needs to have copy argument
https://numpy.org/devdocs/reference/generated/numpy.ndarray.__array__.html
https://github.com/numpy/numpy/issues/25941
