+++ examples/config/phase/mechanical/plastic/dislotwin_IN625.yaml
@@ -2,7 +2,7 @@ type: dislotwin
 
 references:
   - N. Prabhu and M. Diehl,
-    Integrating Materials and Manufacturing Innovation 13:xx-yy, 2008,
+    Integrating Materials and Manufacturing Innovation 13:804-826, 2024,
     https://doi.org/10.1007/s40192-024-00359-1,
     Table 3(d)
   - R. Madec and L.P. Kubin,
