### 2024-11-01 245a83c
false mypy positive (probably because single line)
