### 2024-11-01 #6986 @zhubonan
Hi @s5409693 and @rquey, we are in need of an extra reviewer for this submission, "A Python Library for Pre- and Post-Processing of DAMASK Simulations". Would you be happy to help review it?

The process of reviewing for JOSS is unique and takes place on GitHub. This involves reviewing the manuscript and installing/testing/reading the underlying code. You can learn more about it here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#reviewing-for-joss

Please let me know if you are able/unable to review. Your response is greatly appreciated. Thanks in advance.
