### 2024-11-04 #6986 @s5409693
Hi Bonan Zhu,

I am honored to be considered as a JOSS reviewer, but unfortunately, I’m currently not able to provide a thorough review, so I have to decline. Sorry about that.

Best,

Paul

From: Bonan Zhu ***@***.***>
Sent: Friday, November 1, 2024 3:25 PM
Cc: Seibert, Paul ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: A Python Library for Pre- and Post-Processing of DAMASK Simulations (Issue #6986)

Hi @s5409693 <https://github.com/s5409693>  and @rquey <https://github.com/rquey> , we are in need of an extra reviewer for this submission, "A Python Library for Pre- and Post-Processing of DAMASK Simulations". Would you be happy to help review it?

The process of reviewing for JOSS is unique and takes place on GitHub. This involves reviewing the manuscript and installing/testing/reading the underlying code. You can learn more about it here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#reviewing-for-joss

Please let me know if you are able/unable to review. Your response is greatly appreciated. Thanks in advance.
