### 2024-11-05 1dbaa9c
more direct logic
