### 2024-11-13 13b45e2
note on how to calculate actual shape change from kinematics
