### 2024-11-13 fb0f6e3
put init calls close to each other
