### 2024-11-18 #6986 @MarDiehl
some further suggestions
- mgroeber
- imikejackson
- drowenhorst-nrl

and active developers from https://github.com/mtex-toolbox/mtex
