### 2024-11-19 #7164 @mkuehbach
- [ ] P1L23: "workflows"
- [ ] P1L31: and and construct, "thus allowing for"
- [ ] P1L32: "as well" here suggests that there is a priority of the dependencies of that is likely not so strong and eventually "and" should be used instead of "as well as"
- [ ] P1L37: "manipulation", that term sometimes sets a negative tone, given that the tool works with programmatic instructions, hence "instantiation" could be better or modification if one is to modify a previously generated structure
- [ ] P2L42: "are based on a"
- [ ] P3L64: FAIR in which regards, i.e. check against e.g. Jacobsen 2020 and Wilkinson 2016 papers and better state specifically which points i.e. F1, I2 etc. are addressed. The statement the authors make is not incorrect but take e.g. the point on metadata concepts used in DADF5. Is there an API to query these concepts, if not e.g. the point on FAIR is not met. Therefore, it is typically more informative to be here more specific. This can be done in a single sentence.
