### 2024-11-20 d50f911
bugfix: wrong order for differential operators
need to take 'F' order into account. Outsourced into tested functions
