### 2024-11-21 e708a21
improved naming, new option to flatten
