+++ .gitlab-ci.yml
@@ -268,13 +268,21 @@ grid_performance:
 
 
 ###################################################################################################
-update_website:
+update_website_master:
   stage: finalize
   trigger:
     project: damask/website
     branch: master
   only:
     - development
+
+update_website_3.0:
+  stage: finalize
+  trigger:
+    project: damask/website
+    branch: 3.0
+  only:
+    - 3.0
 
 update_revision:
   stage: finalize
