### 2024-12-10 6f905bd
consistent reshaping tuple syntax
