### 2024-12-12 28d7350
Merge branch '466-python-package-divergence-p-is-high' into 'development'
bugfix: wrong order for differential operators

Closes #466

See merge request damask/DAMASK!999
