### 2024-12-12 #7164 @MarDiehl
> @MarDiehl I have you made the changes as suggested by @mkuehbach?

yes, they are included in the updated proof.
