### 2024-12-12 f3db604
Docker ci gnu intel
