### 2024-12-12 db6623f
bump minimum version requirements
- Python: 3.9 -> 3.10
- CMake: 3.12 -> 3.18
- GCC/gfortran: 9.1 -> 11.1
