+++ python/setup.cfg
@@ -17,7 +17,7 @@ classifiers =
 [options]
 packages = find:
 include_package_data = true
-python_requires = >= 3.9
+python_requires = >= 3.10
 install_requires =
     pandas>=0.24                                                                                    # requires numpy
     numpy>=1.17                                                                                     # needed for default_rng
