### 2024-12-12 bf84494
require newer Python versions
minimum according to Ubuntu 22.04
