+++ python/setup.cfg
@@ -19,10 +19,10 @@ packages = find:
 include_package_data = true
 python_requires = >= 3.10
 install_requires =
-    pandas>=0.24                                                                                    # requires numpy
-    numpy>=1.17                                                                                     # needed for default_rng
-    scipy>=1.2
-    h5py>=2.9                                                                                       # requires numpy
-    vtk>=8.1
-    matplotlib>=3.0                                                                                 # requires numpy, pillow
-    pyyaml>=3.12
+    pandas>=1.3
+    numpy>=1.21
+    scipy>=1.8
+    h5py>=3.6
+    vtk>=9.1
+    matplotlib>=3.5
+    pyyaml>=5.3
