### 2024-12-13 d63af52
modern python build via pyproject.toml only
