+++ python/setup.cfg
@@ -1,28 +1,0 @@
-[metadata]
-name = damask
-version = file: damask/VERSION
-author = The DAMASK team
-author_email = support@damask-multiphyiscs.org
-url = https://damask-multiphysics.org
-description = DAMASK processing tools
-long_description = Pre- and post-processing tools for DAMASK
-license = AGPL3
-classifiers =
-    Intended Audience :: Science/Research
-    Topic :: Scientific/Engineering
-    Programming Language :: Python :: 3
-    License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)
-    Operating System :: OS Independent
-
-[options]
-packages = find:
-include_package_data = true
-python_requires = >= 3.10
-install_requires =
-    pandas>=1.3
-    numpy>=1.21
-    scipy>=1.8
-    h5py>=3.6
-    vtk>=9.1
-    matplotlib>=3.5
-    pyyaml>=5.3
