### 2024-12-13 8e0695e
workarounds not needed anymore
- pyyaml: https://stackoverflow.com/questions/16782112
- numpy: https://numpy.org/doc/2.0/release/1.19.0-notes.html#numpy-random-bit-generator-moved-to-numpy-random-bit-generator
