### 2024-12-13 b958924
avoid code duplication
