+++ .gitlab-ci.yml
@@ -87,6 +87,18 @@ pytest:
     - python3 -m pytest -v --cov=damask --cov-report=xml --damaskroot=${DAMASK_ROOT}
     - coverage report --fail-under=90 --show-missing
 
+doctest:
+  stage: python
+  tags:
+    - matesting2-container-runner
+  image: git.damask-multiphysics.org:5050/damask/damask:python310_2407
+  before_script: []
+  script:
+    - export DAMASK_ROOT=$(pwd)
+    - PYTHONPATH=${DAMASK_ROOT}/python:$PYTHONPATH
+    - cd python
+    - python3 -m pytest -v --doctest-modules -k 'not result and not configmaterial and not orientation and not rotation and not seeds and not geomgrid' damask
+
 mypy:
   stage: python
   tags:
