### 2024-12-16 6bef87e
numpy > 2.0 requires print for nice output
