### 2024-12-17 4a03b8a
track insertion order for base group
this results in correct ordering for increments in HDFView (which are by default
mixed up due to alphanumeric ordering). Note that the status is wrongly
reported:
- https://github.com/HDFGroup/hdfview/issues/347
- https://github.com/HDFGroup/hdf5/issues/5183
