+++ src/CLI.f90
@@ -5,7 +5,7 @@
 !> @author Philip Eisenlohr, Max-Planck-Institut für Eisenforschung GmbH
 !> @brief Parse command line interface for PETSc-based solvers
 !--------------------------------------------------------------------------------------------------
-#define PETSC_MINOR_MIN 12
+#define PETSC_MINOR_MIN 15
 #define PETSC_MINOR_MAX 22
 
 module CLI
