### 2024-12-17 5e886f4
scipy 1.8 is minimum version
https://docs.scipy.org/doc/scipy/reference/generated/scipy.spatial.cKDTree.html#scipy.spatial.cKDTree
