+++ python/damask/_geomgrid.py
@@ -564,12 +564,9 @@ class GeomGrid:
 
         """
         coords = grid_filters.coordinates0_point(cells,size).reshape(-1,3)
-        tree = spatial.cKDTree(seeds,boxsize=size) if periodic else \
-               spatial.cKDTree(seeds)
-        try:
-            material_ = tree.query(coords, workers = int(os.environ.get('OMP_NUM_THREADS',4)))[1]
-        except TypeError:
-            material_ = tree.query(coords, n_jobs = int(os.environ.get('OMP_NUM_THREADS',4)))[1]    # scipy <1.6
+        tree = spatial.KDTree(seeds,boxsize=size) if periodic else \
+               spatial.KDTree(seeds)
+        material_ = tree.query(coords, workers = int(os.environ.get('OMP_NUM_THREADS',4)))[1]
 
         return GeomGrid(material = (material_ if material is None else np.array(material)[material_]).reshape(cells),
                         size     = size,
