### 2024-12-17 676aced
add general metadata to HDF5 file
