+++ CMakeLists.txt
@@ -38,13 +38,31 @@ if(DAMASK_SOLVER_UPPER STREQUAL "GRID")
 endif()
 
 file(STRINGS ${PROJECT_SOURCE_DIR}/VERSION DAMASK_VERSION)
+string(REPLACE "." ";" DAMASK_VERSION_LIST "${DAMASK_VERSION}")
+list(GET DAMASK_VERSION_LIST 0 DAMASK_VERSION_MAJOR)
+list(GET DAMASK_VERSION_LIST 1 DAMASK_VERSION_MINOR)
+list(GET DAMASK_VERSION_LIST 2 DAMASK_VERSION_PATCH)
+string(REPLACE "-" ";" DAMASK_VERSION_PATCH_LIST "${DAMASK_VERSION_PATCH}")
+list(LENGTH DAMASK_VERSION_PATCH_LIST PATCH_CHUNKS)
+if(PATCH_CHUNKS GREATER 1)
+    list(GET DAMASK_VERSION_PATCH_LIST 0 DAMASK_VERSION_PATCH)
+    list(GET DAMASK_VERSION_PATCH_LIST 2 DAMASK_VERSION_HASH)
+endif()
 
 message("\nBuilding ${CMAKE_PROJECT_NAME} ${DAMASK_VERSION}\n")
 
 add_definitions(-DPETSC)
 add_definitions(-DDAMASK_VERSION="${DAMASK_VERSION}")
+add_definitions(-DDAMASK_VERSION_MAJOR="${DAMASK_VERSION_MAJOR}")
+add_definitions(-DDAMASK_VERSION_MINOR="${DAMASK_VERSION_MINOR}")
+add_definitions(-DDAMASK_VERSION_PATCH="${DAMASK_VERSION_PATCH}")
+if(DEFINED DAMASK_VERSION_HASH)
+    add_definitions(-DDAMASK_VERSION_HASH="${DAMASK_VERSION_HASH}")
+endif()
+
 add_definitions(-DCMAKE_SYSTEM_NAME="${CMAKE_SYSTEM_NAME}")
 add_definitions(-DCMAKE_SYSTEM_PROCESSOR="${CMAKE_SYSTEM_PROCESSOR}")
+
 if(PETSC_VERSION VERSION_EQUAL 3.17.0)
    add_definitions("-DCHKERRQ=PetscCall")
 endif()
