### 2024-12-18 bda3874
PETSc 3.15 is minimum supported version
