### 2024-12-18 31c433d
avoid ambiguities
https://english.stackexchange.com/questions/182660
