### 2024-12-18 a14503c
improved error handling/debugging
