+++ src/HDF5_utilities.f90
@@ -122,39 +122,39 @@ subroutine HDF5_utilities_init()
 
 
   call H5Open_f(hdferr)
-  call HDF5_chkerr(hdferr)
+  call HDF5_chkerr(hdferr,__FILE__//':'//IO_intAsStr(__LINE__))
 
   call H5Tget_size_f(H5T_NATIVE_INTEGER,typeSize, hdferr)
-  call HDF5_chkerr(hdferr)
+  call HDF5_chkerr(hdferr,__FILE__//':'//IO_intAsStr(__LINE__))
   if (int(bit_size(0),SIZE_T)/=typeSize*8) &
     error stop 'Default integer size does not match H5T_NATIVE_INTEGER'
 
   call H5Tget_size_f(H5T_NATIVE_DOUBLE,typeSize, hdferr)
-  call HDF5_chkerr(hdferr)
+  call HDF5_chkerr(hdferr,__FILE__//':'//IO_intAsStr(__LINE__))
   if (int(storage_size(0.0_pREAL),SIZE_T)/=typeSize*8) &
     error stop 'pREAL does not match H5T_NATIVE_DOUBLE'
 
   call H5get_libversion_f(HDF5_major,HDF5_minor,HDF5_release,hdferr)
-  call HDF5_chkerr(hdferr)
+  call HDF5_chkerr(hdferr,__FILE__//':'//IO_intAsStr(__LINE__))
   compression_possible = (HDF5_major == 1 .and. HDF5_minor >= 12)                                   ! https://forum.hdfgroup.org/t/6186
 
   call H5Zfilter_avail_f(H5Z_FILTER_DEFLATE_F,avail,hdferr)
-  call HDF5_chkerr(hdferr)
+  call HDF5_chkerr(hdferr,__FILE__//':'//IO_intAsStr(__LINE__))
   compression_possible = compression_possible .and. avail
 
   if (avail) then
     call H5Zget_filter_info_f(H5Z_FILTER_DEFLATE_F,configFlags,hdferr)
-    call HDF5_chkerr(hdferr)
+    call HDF5_chkerr(hdferr,__FILE__//':'//IO_intAsStr(__LINE__))
     compression_possible = compression_possible .and. iand(H5Z_FILTER_ENCODE_ENABLED_F,configFlags) > 0
   end if
 
   call H5Zfilter_avail_f(H5Z_FILTER_SHUFFLE_F,avail,hdferr)
-  call HDF5_chkerr(hdferr)
+  call HDF5_chkerr(hdferr,__FILE__//':'//IO_intAsStr(__LINE__))
   compression_possible = compression_possible .and. avail
 
   if (avail) then
     call H5Zget_filter_info_f(H5Z_FILTER_SHUFFLE_F,configFlags,hdferr)
-    call HDF5_chkerr(hdferr)
+    call HDF5_chkerr(hdferr,__FILE__//':'//IO_intAsStr(__LINE__))
     compression_possible = compression_possible .and. iand(H5Z_FILTER_ENCODE_ENABLED_F,configFlags) > 0
   end if
 
@@ -164,12 +164,18 @@ end subroutine HDF5_utilities_init
 !--------------------------------------------------------------------------------------------------
 !> @brief Check for HDF5 error.
 !--------------------------------------------------------------------------------------------------
-subroutine HDF5_chkerr(e)
+subroutine HDF5_chkerr(e,msg)
 
-  integer, intent(in) :: e
[…]
