### 2024-12-19 d16d965
flexible Miller-Bravais indices
allow to input missing third component, either by using array with
trailing dimension of 3 or by indicating missing components with an
Ellipsis (...).

Also enabled use of integers represented as floats in conversion
function and introduced MillerBravais as alias for Bravais.
