### 2024-12-20 e79d273
language polish and keepdims bugfix
