### 2024-12-23 ceea7e5
improved reporting
CMAKE actually reports the target system
