+++ CMakeLists.txt
@@ -40,8 +40,9 @@ file(STRINGS ${PROJECT_SOURCE_DIR}/VERSION DAMASK_VERSION)
 message("\nBuilding ${CMAKE_PROJECT_NAME} ${DAMASK_VERSION}\n")
 
 add_definitions(-DPETSC)
-add_definitions(-DDAMASKVERSION="${DAMASK_VERSION}")
-add_definitions(-DCMAKE_SYSTEM="${CMAKE_SYSTEM}")
+add_definitions(-DDAMASK_VERSION="${DAMASK_VERSION}")
+add_definitions(-DCMAKE_SYSTEM_NAME="${CMAKE_SYSTEM_NAME}")
+add_definitions(-DCMAKE_SYSTEM_PROCESSOR="${CMAKE_SYSTEM_PROCESSOR}")
 if(PETSC_VERSION VERSION_EQUAL 3.17.0)
    add_definitions("-DCHKERRQ=PetscCall")
 endif()
