### 2024-12-23 260d88c
clearer logic
