+++ src/grid/grid_mech_spectral_Galerkin.f90
@@ -308,8 +308,7 @@ subroutine grid_mechanical_spectral_Galerkin_init(num_grid)
     F = reshape(temp33n,[9,cells(1),cells(2),cells3])
     call HDF5_read(temp33n,groupHandle,'F_lastInc')
     F_lastInc = reshape(temp33n,[3,3,cells(1),cells(2),cells3])
-
-  elseif (CLI_restartInc == 0) then restartRead
+  else restartRead
     F_lastInc = spread(spread(spread(math_I3,3,cells(1)),4,cells(2)),5,cells3)                      ! initialize to identity
     F = reshape(F_lastInc,[9,cells(1),cells(2),cells3])
   end if restartRead
@@ -335,7 +334,6 @@ subroutine grid_mechanical_spectral_Galerkin_init(num_grid)
 
     call HDF5_closeGroup(groupHandle)
     call HDF5_closeFile(fileHandle)
-
   end if restartRead2
 
 end subroutine grid_mechanical_spectral_Galerkin_init
