+++ src/grid/grid_mech_spectral_basic.f90
@@ -220,8 +220,7 @@ subroutine grid_mechanical_spectral_basic_init(num_grid)
     F = reshape(temp33n,[9,cells(1),cells(2),cells3])
     call HDF5_read(temp33n,groupHandle,'F_lastInc')
     F_lastInc = reshape(temp33n,[3,3,cells(1),cells(2),cells3])
-
-  elseif (CLI_restartInc == 0) then restartRead
+  else restartRead
     F_lastInc = spread(spread(spread(math_I3,3,cells(1)),4,cells(2)),5,cells3)                      ! initialize to identity
     F = reshape(F_lastInc,[9,cells(1),cells(2),cells3])
   end if restartRead
@@ -247,7 +246,6 @@ subroutine grid_mechanical_spectral_basic_init(num_grid)
 
     call HDF5_closeGroup(groupHandle)
     call HDF5_closeFile(fileHandle)
-
   end if restartRead2
 
   call utilities_updateGamma(C_minMaxAvg)
