### 2025-01-02 #7164 @zhubonan
Hi @MarDiehl, I am happy proceed with the acceptance of the paper as two reviewers have recommended publication. Sorry for the delay, as I was thinking it would be better to double-check with @mkuehbach.

You will see a list of things for the authors to do in the generated 'post-review checklist'. It mainly involves making a release, archive the code on Zenodo, let me know its DOI.

When archiving the code on Zenodo, please make sure the title and authors matches with that of the paper. It would be useful to also include an abstract.
Note that if you link GitHub with Zenodo then the default title will be the repository's one which needs to be changed.
