### 2025-01-02 c44b998
typos etc spotted by Philip
