### 2025-01-04 14fb815
zlib is only needed for grid solver (VTI)
