### 2025-01-06 07610a4
Debug slow pipeline
