+++ .gitlab-ci.yml
@@ -59,7 +59,7 @@ setuptools:
   stage: python
   tags:
     - matesting2-container-runner
-  image: git.damask-multiphysics.org:5050/damask/damask:python310_2407
+  image: git.damask-multiphysics.org:5050/damask/damask:python310_2412
   before_script: []
   script:
     - sed -i 's/-[[:digit:]]*-.*//' VERSION
@@ -70,7 +70,7 @@ pytest:
   stage: python
   tags:
     - matesting2-container-runner
-  image: git.damask-multiphysics.org:5050/damask/damask:python310_2407
+  image: git.damask-multiphysics.org:5050/damask/damask:python310_2412
   before_script: []
   script:
     - export DAMASK_ROOT=$(pwd)
@@ -83,7 +83,7 @@ mypy:
   stage: python
   tags:
     - matesting2-container-runner
-  image: git.damask-multiphysics.org:5050/damask/damask:python310_2407
+  image: git.damask-multiphysics.org:5050/damask/damask:python310_2412
   before_script: []
   script:
     - cd python
@@ -95,7 +95,7 @@ unittest_GNU_DEBUG:
   stage: compile
   tags:
     - matesting2-container-runner
-  image: git.damask-multiphysics.org:5050/damask/damask:ubuntu2404_gnu13_petsc2106
+  image: git.damask-multiphysics.org:5050/damask/damask:ubuntu2404_gnu13_petsc2106_ml
   before_script:
     - TEMPDIR=$(mktemp -d /tmp/tmp.XXXXXXXXXX)
     - cmake -B ${TEMPDIR} -DDAMASK_SOLVER=test -DCMAKE_INSTALL_PREFIX=${TEMPDIR} -DCMAKE_BUILD_TYPE=RELEASE -DBUILDCMD_POST=-coverage
@@ -109,7 +109,7 @@ unittest_GNU_RELEASE:
   stage: compile
   tags:
     - matesting2-container-runner
-  image: git.damask-multiphysics.org:5050/damask/damask:ubuntu2404_gnu13_petsc2106
+  image: git.damask-multiphysics.org:5050/damask/damask:ubuntu2404_gnu13_petsc2106_ml
   before_script:
     - TEMPDIR=$(mktemp -d /tmp/tmp.XXXXXXXXXX)
     - cmake -B ${TEMPDIR} -DDAMASK_SOLVER=test -DCMAKE_INSTALL_PREFIX=${TEMPDIR} -DCMAKE_BUILD_TYPE=RELEASE -DBUILDCMD_POST=-coverage
@@ -123,7 +123,7 @@ unittest_GNU_PERFORMANCE:
   stage: compile
   tags:
     - matesting2-container-runner
-  image: git.damask-multiphysics.org:5050/damask/damask:ubuntu2404_gnu13_petsc2106
+  image: git.damask-multiphysics.org:5050/damask/damask:ubuntu2404_gnu13_petsc2106_ml
   before_script:
     - TEMPDIR=$(mktemp -d /tmp/tmp.XXXXXXXXXX)
     - cmake -B ${TEMPDIR} -DDAMASK_SOLVER=test -DCMAKE_INSTALL_PREFIX=${TEMPDIR} -DCMAKE_BUILD_TYPE=PERFORMANCE -DBUILDCMD_POST=-coverage
@@ -137,7 +137,7 @@ grid_GNU:
   stage: compile
   tags:
     - matesting2-container-runner
-  image: git.damask-multiphysics.org:5050/damask/damask:ubuntu2404_gnu13_petsc2106
+  image: git.damask-multiphysics.org:5050/damask/damask:ubuntu2404_gnu13_petsc2106_ml
   before_script:
[…]
