+++ paper.md
@@ -22,7 +22,7 @@ authors:
   corresponding: true
 
 affiliations:
-- name: Max-Planck-Institute for Sustainable Materials, 40237 Düsseldorf, Germany
+- name: Max Planck Institute for Sustainable Materials, 40237 Düsseldorf, Germany
   index: 1
 - name: Chemical Engineering and Materials Science, Michigan State University, East Lansing, MI 48824, USA
   index: 2
