### 2025-01-06 #7164 @MarDiehl
@zhubonan
great news!

The current version of the paper (incl. ORCID) is checked and agreed on by all authors.
The latest release of DAMASK is 3.0.1, it's on Zenodo with DOI [10.5281/zenodo.14586846](https://doi.org/10.5281/zenodo.14586846). The general DOI (for all releases on Zenodo) is https://doi.org/10.5281/zenodo.14586837. I would prefer to use the latter one if that's ok with JOSS guidelines. The list of authors on Zenodo is for the full DAMASK release (not only for the Python tools) and, hence, longer.

@eisenlohr @sharanroongta @dmentock @FranzRoters
