+++ paper.bib
@@ -1,6 +1,6 @@
 @article{roters2019damask,
-  title={{DAMASK}--The {D}{\"u}sseldorf {A}dvanced {M}aterial {S}imulation {K}it for modeling multi-physics crystal plasticity, thermal, and damage phenomena from the single crystal up to the component scale},
-  author={Roters, Franz and Diehl, Martin and Shanthraj, Pratheek and Eisenlohr, Philip and Reuber, C and Wong, Su Leen and Maiti, Tias and Ebrahimi, Alireza and Hochrainer, Thomas and Fabritius, H-O and others},
+  title={{DAMASK} -- The {D}{\"u}sseldorf {A}dvanced {M}aterial {S}imulation {K}it for modeling multi-physics crystal plasticity, thermal, and damage phenomena from the single crystal up to the component scale},
+  author= {Roters, F. and Diehl, M. and Shanthraj, P. and Eisenlohr, P. and Reuber, C. and Wong, S. L. and Maiti, T. and Ebrahimi, A. and Hochrainer, T. and Fabritius, H.-O. and Nikolov, S. and Friak, M. and Fujita, N. and Grilli, N. and Janssens, K. G. F. and Jia, N. and Kok, P. J. J. and Ma, D. and Meier, F. and Werner, E. and Stricker, M. and Weygand, D. and Raabe, D.},
   journal={Computational Materials Science},
   volume={158},
   pages={420--478},
@@ -74,12 +74,12 @@
 @Article{         harris2020array,
  title         = {Array programming with {NumPy}},
  author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
-                 van der Walt and Ralf Gommers and Pauli Virtanen and David
+         {van der Walt} and Ralf Gommers and Pauli Virtanen and David
                  Cournapeau and Eric Wieser and Julian Taylor and Sebastian
                  Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
                  and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
-                 Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
-                 R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
+                 Brett and Allan Haldane and Jaime Fern{\'{a}}ndez {del
+                 R{\'{i}}o} and Mark Wiebe and Pearu Peterson and Pierre
                  G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
                  Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
                  Travis E. Oliphant},
@@ -95,7 +95,7 @@
 }
 
 @Article{WilkinsonEtAl2016,
[…]
