### 2025-01-07 6caf6cd
reporting to debug sporadic test failures
uses logic from PRIVATE/testing/conftest.py
