### 2025-01-07 9120e1b
centralize configuration
