### 2025-01-07 #7164 @zhubonan
@MarDiehl Looks like there are two duplicated/unused cite entries:

https://github.com/damask-multiphysics/DAMASK/blob/5b95ccb2cacca6a1eb75b51517f4b620dfcabff5/paper.bib#L118C1-L128C1

https://github.com/damask-multiphysics/DAMASK/blob/5b95ccb2cacca6a1eb75b51517f4b620dfcabff5/paper.bib#L23-L32

If you remove them from the paper.bib the warning should go away.
