+++ paper.bib
@@ -20,17 +20,6 @@
   doi = {10.1186/2193-9772-3-5}
 }
 
-@article{nolze2016orientations,
-  title={Orientations--perfectly colored},
-  author={Nolze, Gert and Hielscher, R},
-  journal={Journal of Applied Crystallography},
-  volume={49},
-  number={5},
-  pages={1786--1802},
-  year={2016},
-  publisher={International Union of Crystallography}
-}
-
 @article{johnstone2020density,
   title={Density-based clustering of crystal (mis)orientations and the orix Python library},
   author={Johnstone, Duncan N and Martineau, Ben H and Crout, Phillip and Midgley, Paul A and Eggeman, Alexander S},
@@ -104,26 +93,6 @@
   file         = {:2016/WilkinsonEtAl2016 - The FAIR Guiding Principles for Scientific Data Management and Stewardship.pdf:PDF},
   journaltitle = {Scientific Data},
   publisher    = {Springer Science and Business Media {LLC}},
-}
-
-
-@misc{YAML,
-  author = {{The YAML development team}},
-  title = {YAML},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/yaml/www.yaml.org}
-}
-
-@article{quey2011large,
-  title={Large-scale 3D random polycrystals for the finite element method: Generation, meshing and remeshing},
-  author={Quey, Romain and Dawson, PR and Barbe, Fabrice},
-  journal={Computer Methods in Applied Mechanics and Engineering},
-  volume={200},
-  number={17-20},
-  pages={1729--1745},
-  year={2011},
-  publisher={Elsevier}
 }
 
 @Inbook{ahrens200536,
