### 2025-01-07 #7164 @zhubonan
BTW The Zenodo archive's author list is not the same as that in the paper. I guess it is because the archive is for the entire DAMASK code base and the paper is for the python library?
