### 2025-01-08 798f390
additional version information for Marc simulations
