### 2025-01-09 7a1c397
tagging jobs to be run on bare metal
also centralized specification of compile job
