+++ .gitlab-ci.yml
@@ -109,7 +109,7 @@ mypy:
     - ./bin/DAMASK_test
     - find . -name '*.gcda' -not -path "**/test/*" -exec gcov {} \;
 
-#--------------------------------------------------------------------------------------------------
+# -------------------------------------------------------------------------------------------------
 
 .unittest_GNU:
   extends: .unittest
@@ -136,7 +136,7 @@ unittest_GNU_PERFORMANCE:
     - cmake -B ${TEMPDIR} -DDAMASK_SOLVER=test -DCMAKE_INSTALL_PREFIX=${TEMPDIR} -DCMAKE_BUILD_TYPE=PERFORMANCE -DBUILDCMD_POST=-coverage
     - cmake --build ${TEMPDIR} --target install
 
-#--------------------------------------------------------------------------------------------------
+# -------------------------------------------------------------------------------------------------
 
 .unittest_Intel:
   extends: .unittest
@@ -171,76 +171,66 @@ unittest_Intel_PERFORMANCE:
 
 ###################################################################################################
 
-grid_GNU:
+.compile_grid:
   stage: compile
   tags:
     - matesting2-container-runner
-  image: git.damask-multiphysics.org:5050/damask/damask:ubuntu2404_gnu13_petsc2106_ml
-  before_script:
-    - source env/DAMASK.sh
   script:
     - cd PRIVATE/testing
     - pytest -k 'compile and grid'
+
+.compile_mesh:
+  stage: compile
+  tags:
+    - matesting2-container-runner
+  script:
+    - cd PRIVATE/testing
+    - pytest -k 'compile and mesh'
+
+grid_GNU:
+  extends: .compile_grid
+  image: git.damask-multiphysics.org:5050/damask/damask:ubuntu2404_gnu13_petsc2106_ml
+  before_script:
+    - source env/DAMASK.sh
 
 mesh_GNU:
-  stage: compile
-  tags:
-    - matesting2-container-runner
+  extends: .compile_mesh
   image: git.damask-multiphysics.org:5050/damask/damask:ubuntu2404_gnu13_petsc2106_ml
   before_script:
     - source env/DAMASK.sh
-  script:
-    - cd PRIVATE/testing
-    - pytest -k 'compile and mesh'
 
 grid_GNU-64bit:
-  stage: compile
-  tags:
-    - matesting2-container-runner
+  extends: .compile_grid
   image: git.damask-multiphysics.org:5050/damask/damask:ubuntu2404_gnu13_petsc2106_int64
   before_script:
     - source env/DAMASK.sh
-  script:
-    - cd PRIVATE/testing
-    - pytest -k 'compile and grid'
 
 mesh_GNU-64bit:
-  stage: compile
+  extends: .compile_mesh
   tags:
     - matesting2-container-runner
   image: git.damask-multiphysics.org:5050/damask/damask:ubuntu2404_gnu13_petsc2106_int64
   before_script:
     - source env/DAMASK.sh
-  script:
-    - cd PRIVATE/testing
-    - pytest -k 'compile and mesh'
 
-grid_IntelLLVM:
-  stage: compile
-  tags:
-    - matesting2-container-runner
+grid_Intel:
+  extends: .compile_grid
   image: git.damask-multiphysics.org:5050/damask/damask:ubuntu2404_intelllvm20250_petsc2106_ml
   before_script:
     - source /opt/intel/oneapi/setvars.sh
     - source env/DAMASK.sh
-  script:
-    - cd PRIVATE/testing
-    - pytest -k 'compile and grid'
 
-mesh_IntelLLVM:
[…]
