### 2025-01-11 16f75c8
"Schmid" can be defined in Crystal
some reshaping is needed when inheriting in "Rotation", might be
changed in a 4.0 release if this would increse consistency
