### 2025-01-11 c4a64fd
control and report random state for tests
Entropy used for seeding is reported, this allows--in combination with
the new CLI flag to set it--to re-run failed tests.

Some calls to np.random.XXX still exists in pytest.parametrize
decorators. Fixing them is not straight forward
