+++ python/tests/test_Rotation.py
@@ -21,10 +21,10 @@ def set_of_rotations(set_of_quaternions):
     return [Rotation.from_quaternion(s) for s in set_of_quaternions]
 
 @pytest.fixture
-def multidim_rotations(set_of_quaternions):
+def multidim_rotations(np_rng,set_of_quaternions):
     L = len(set_of_quaternions)
     i = 0
-    while L%(f:=np.random.randint(2,np.sqrt(L).astype(int))) > 0 and i<L:
+    while L%(f:=np_rng.integers(2,np.sqrt(L).astype(int))) > 0 and i<L:
         i += 1
 
     f = i if i == L else f
@@ -689,9 +689,9 @@ class TestRotation:
             assert np.allclose(single(u),c) and np.allclose(single(u),vectorized(u)), f'{u},{c}'
 
     @pytest.mark.parametrize('func',[Rotation.from_axis_angle])
-    def test_normalization_vectorization(self,func):
+    def test_normalization_vectorization(self,np_rng,func):
         """Check vectorized implementation normalization."""
-        vec = np.random.rand(5,4)
+        vec = np_rng.random((5,4))
         ori = func(vec,normalize=True)
         for v,o in zip(vec,ori):
             assert np.allclose(func(v,normalize=True).as_quaternion(),o.as_quaternion())
@@ -700,8 +700,8 @@ class TestRotation:
         with pytest.raises(TypeError):
             Rotation(np.ones(3))
 
-    def test_to_numpy(self):
-        r = Rotation.from_random(np.random.randint(0,10,4))
+    def test_to_numpy(self,np_rng):
+        r = Rotation.from_random(np_rng.integers(0,10,4),rng_seed=np_rng)
         assert (r.as_quaternion() == np.array(r)).all()
 
     def test_bounds(self,multidim_rotations):
@@ -859,28 +859,28 @@ class TestRotation:
 
 
     @pytest.mark.parametrize('shape',[None,1,(4,4)])
-    def test_random(self,shape):
-        r = Rotation.from_random(shape)
+    def test_random(self,np_rng,shape):
+        r = Rotation.from_random(shape,rng_seed=np_rng)
         assert r.shape == () if shape is None else (1,) if shape == 1 else shape
 
     @pytest.mark.parametrize('shape',[None,5,(4,6)])
-    def test_equal(self,shape):
-        R = Rotation.from_random(shape,rng_seed=1)
+    def test_equal(self,np_rng,shape):
+        R = Rotation.from_random(shape,rng_seed=np_rng)
         assert R == R if shape is None else (R == R).all()
 
     @pytest.mark.parametrize('shape',[None,5,(4,6)])
-    def test_allclose(self,shape):
-        R = Rotation.from_random(shape,rng_seed=1)
+    def test_allclose(self,np_rng,shape):
+        R = Rotation.from_random(shape,rng_seed=np_rng)
         assert R.allclose(R)
 
     @pytest.mark.parametrize('shape',[None,5,(4,6)])
-    def test_unequal(self,shape):
-        R = Rotation.from_random(shape,rng_seed=1)
+    def test_unequal(self,np_rng,shape):
+        R = Rotation.from_random(shape,rng_seed=np_rng)
         assert not (R != R if shape is None else (R != R).any())
 
 
-    def test_equal_ambiguous(self):
-        qu = np.random.rand(10,4)
+    def test_equal_ambiguous(self,np_rng):
+        qu = np_rng.random((10,4))
         qu[:,0] = 0.
         qu /= np.linalg.norm(qu,axis=1,keepdims=True)
         assert (Rotation(qu) == Rotation(-qu)).all()
@@ -889,26 +889,26 @@ class TestRotation:
         assert (multidim_rotations == ~~multidim_rotations).all()
 
[…]

[…]
