### 2025-01-13 51d43de
better consistency checks for thermal and damage
if thermal/damage are active, T/phi are defined in the whole domain
which requires to have C_p,rho,K/mu,l_c defined in each phase.
