### 2025-01-13 a4c661b
detect invalid type for damage
