+++ src/phase_mechanical_elastic.f90
@@ -46,7 +46,8 @@ module subroutine elastic_init(phases)
     print'(/,1x,a,1x,i0,a)', 'phase',ph,': '//phases%key(ph)
     refs = config_listReferences(elastic,indent=3)
     if (len(refs) > 0) print'(/,1x,a)', refs
-    if (elastic%get_asStr('type') /= 'Hooke') call IO_error(200,ext_msg=elastic%get_asStr('type'))
+    if (elastic%get_asStr('type') /= 'Hooke') &
+      call IO_error(200,label1='elasticity',ext_msg=elastic%get_asStr('type'))
 
     associate(prm => param(ph))
 
