### 2025-01-13 8be41d0
suggestion to prevent uninteded behaviour for scalar Rotations and Orientations
