### 2025-01-13 2f10d36
from_random without shape argument should give scalar
this is consistent to from_quaternion etc.
There is some ambiguity on how to handle shape=1. Currently we have
shape=1 equivalent to shape=None
Alternatively, we could have shape=1 equivalent to shape=[1].
