+++ python/damask/_rotation.py
@@ -156,7 +156,6 @@ class Rotation:
                np.logical_or(np.all(self.quaternion ==     other.quaternion,axis=-1),
                              np.all(self.quaternion == -1.*other.quaternion,axis=-1))
 
-
     def __ne__(self,
                other: object) -> bool:
         """
@@ -170,13 +169,14 @@ class Rotation:
             Rotation to check for inequality.
 
         """
-        return np.logical_not(self==other) if isinstance(other, Rotation) else NotImplemented
+        return np.logical_not(self==other)
+
 
     def isclose(self: MyType,
                 other: MyType,
                 rtol: float = 1.e-5,
                 atol: float = 1.e-8,
-                equal_nan: bool = True) -> bool:
+                equal_nan: bool = True) -> np.ndarray:
         """
         Report where values are approximately equal to corresponding ones of other Rotation.
 
@@ -202,12 +202,11 @@ class Rotation:
         return np.logical_or(np.all(np.isclose(s,    o,rtol,atol,equal_nan),axis=-1),
                              np.all(np.isclose(s,-1.*o,rtol,atol,equal_nan),axis=-1))
 
-
     def allclose(self: MyType,
                  other: MyType,
                  rtol: float = 1.e-5,
                  atol: float = 1.e-8,
-                 equal_nan: bool = True) -> Union[np.bool_, bool]:
+                 equal_nan: bool = True) -> np.bool_:
         """
         Test whether all values are approximately equal to corresponding ones of other Rotation.
 
