### 2025-01-16 41a106b
correct type annotation
Rotation/Orientation follow NumPy and return an array of booleans. This
violates Liskovs substitution principle and is therefore causing an
error in mypy
