### 2025-01-16 49e914d
use proper logical operations, not bit-wise operators
