### 2025-01-17 f051135
Temp. dep. K and Cp
