### 2025-01-17 e489a1e
use "C" for crystal object; introduce shorthand variable
