### 2025-01-18 368cad8
manual definition of arguments
autocompletion for arguments works now. Also, docstring extension (at
least the test) was broken in Python 3.13 because whitespaces are
stripped (https://docs.python.org/3/whatsnew/3.13.html).

Given that our API is rather stable, further changes and the associated
efforts to keep Rotation/Crystal and Orientation in sync are hopefully
not to frequent.

Explicit definition also revealed a bug in "from_parallel" where "a" and "b"
had two meanings.
