+++ python/damask/util.py
@@ -9,10 +9,9 @@ import re as _re
 import signal as _signal
 import fractions as _fractions
 import contextlib as _contextlib
-from collections import abc as _abc, OrderedDict as _OrderedDict
-from functools import reduce as _reduce, partial as _partial, wraps as _wraps
-import inspect
-from typing import Optional as _Optional, Callable as _Callable, Union as _Union, Iterable as _Iterable, \
+from collections import abc as _abc
+from functools import reduce as _reduce, partial as _partial
+from typing import Optional as _Optional, Union as _Union, Iterable as _Iterable, \
                    Dict as _Dict, List as _List, Tuple as _Tuple, Literal as _Literal, \
                    Any as _Any, TextIO as _TextIO, Generator as _Generator
 from pathlib import Path as _Path
@@ -455,7 +454,7 @@ def hybrid_IA(dist: _FloatSequence,
 
     """
     N_opt_samples = max(_np.count_nonzero(dist),N)                                                  # random subsampling if too little samples requested
-    N_inv_samples = 0
+    N_inv_samples = _np.int_(0)
 
     scale_,scale,inc_factor = (0.0,float(N_opt_samples),1.0)
     while (not _np.isclose(scale, scale_)) and (N_inv_samples != N_opt_samples):
@@ -579,164 +578,6 @@ def shapeblender(a: _Tuple[int, ...],
             _b = ((1,) if len(_b)<len(a_) else ()) + _b
         return _np.broadcast_shapes(a_,_b)
 
-
-def _docstringer(docstring: _Union[str, _Callable],
-                 adopted_parameters: _Union[None, str, _Callable] = None,
-                 adopted_return: _Union[None, str, _Callable] = None,
-                 adopted_notes: _Union[None, str, _Callable] = None,
-                 adopted_examples: _Union[None, str, _Callable] = None,
-                 adopted_references: _Union[None, str, _Callable] = None) -> str:
-    """
-    Extend a docstring.
-
-    Parameters
-    ----------
-    docstring : str or callable, optional
-       Docstring (of callable) to extend.
-    adopted_* : str or callable, optional
-       Additional information to insert into or append to respective section.
-
-    Notes
-    -----
-    adopted_return fetches the typehint of a passed function instead of the docstring
-
-    """
-    docstring_: str = str(     docstring if isinstance(docstring,str)
-                          else docstring.__doc__ if callable(docstring) and docstring.__doc__
-                          else '').rstrip()+'\n'
-    sections = _OrderedDict(
-        Parameters=adopted_parameters,
-        Returns=adopted_return,
-        Examples=adopted_examples,
-        Notes=adopted_notes,
-        References=adopted_references)
-
-    for i, (key, adopted) in [(i,(k,v)) for (i,(k,v)) in enumerate(sections.items()) if v is not None]:
-        section_regex = fr'^([ ]*){key}\s*\n\1*{"-"*len(key)}\s*\n'
-        if key=='Returns':
-            if callable(adopted):
-                return_class = adopted.__annotations__.get('return','')
-                return_type_ = (_sys.modules[adopted.__module__].__name__.split('.')[0]
-                                +'.'
-                                +(return_class.__name__ if not isinstance(return_class,str) else return_class))
-            else:
-                return_type_ = adopted
-            docstring_ = _re.sub(fr'(^[ ]*{key}\s*\n\s*{"-"*len(key)}\s*\n[ ]*[A-Za-z0-9_ ]*: )(.*)\n',
-                                 fr'\1{return_type_}\n',
-                                 docstring_,flags=_re.MULTILINE)
-        else:
-            section_content_regex = fr'{section_regex}(?P<content>.*?)\n *(\n|\Z)'
[…]

[…]
