### 2025-01-19 c558bf3
improved typehints
