+++ python/damask/_table.py
@@ -421,7 +421,7 @@ class Table:
             if m:
                 idx = np.ravel_multi_index(tuple(map(int,m.group(2).split(","))),
                                            self.shapes[key])
-                iloc = dup.data.columns.get_loc(key).tolist().index(True) + idx
+                iloc = dup.data.columns.get_loc(key).tolist().index(True) + idx                     # type: ignore
                 dup.data.iloc[:,iloc] = data
             else:
                 dup.data[label]       = data.reshape(dup.data[label].shape)
