### 2025-01-20 3155dd6
include scientific notation in regex
