### 2025-01-21 #7164 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
