### 2025-01-21 #7164 @MarDiehl
Just spotted a small glitch:
The surname of the first author is "Otto de Mentock", given name is "Daniel".
