### 2025-01-21 #7164 @MarDiehl
In the Markdown file we have `name: Daniel Otto de Mentock`, can we differentiate surname and given name there?
