### 2025-01-21 5eaaa77
don't rely on autodetection for name split
