### 2025-01-21 #7164 @MarDiehl
@dmentock: can you confirm that the footer in the latest proof is correct?
