### 2025-01-22 76bca50
logging allows users to conveniently silence info output when running scripts with the library
newlines between import statements and code should be consistent
