+++ python/damask/_result.py
@@ -8,6 +8,7 @@ import functools
 from pathlib import Path
 from collections import defaultdict
 from collections.abc import Iterable
+import logging
 from typing import Optional, Union, Callable, Any, Sequence, Literal, Dict, List, Tuple
 
 import h5py
@@ -26,10 +27,11 @@ from . import util
 from ._typehints import FloatSequence, IntSequence, DADF5Dataset
 
 
+logger = logging.getLogger(__name__)
+
 chunk_size = 1024**2//8                                                                             # for compression in HDF5
 
 prefix_inc = 'increment_'
-
 
 def _read(dataset: h5py._hl.dataset.Dataset) -> np.ndarray:
     """Read a dataset and its metadata into a numpy.ndarray."""
@@ -352,7 +354,7 @@ class Result:
         dup = self._manage_view('set',increments,times,phases,homogenizations,fields)
         if protected is not None:
             if not protected:
-                print(util.warn('Warning: Modification of existing datasets allowed!'))
+                logger.warning(util.warn('Modification of existing datasets allowed!'))
             dup._protected = protected
 
         return dup
@@ -566,7 +568,7 @@ class Result:
     def enable_user_function(self,
                              func: Callable):
         globals()[func.__name__]=func
-        print(f'Function {func.__name__} enabled in add_calculation.')
+        logger.info(f'Function {func.__name__} enabled in add_calculation.')
 
 
     @property
@@ -1528,7 +1530,7 @@ class Result:
                                           'meta': {k: v for k,v in loc.attrs.items()}}
                 return callback(**datasets_in,**args)
             except Exception as err:
-                print(f'Error during calculation: {err}.')
+                logger.error(f'Error during pointwise calculation: {err}.')
                 return None
 
         groups = []
@@ -1541,7 +1543,7 @@ class Result:
                             if set(datasets.values()).issubset(f[group].keys()): groups.append(group)
 
         if len(groups) == 0:
-            print('No matching dataset found, no data was added.')
+            logger.warning('No matching dataset found, no data was added.')
             return
 
 
@@ -1574,7 +1576,7 @@ class Result:
                     dataset.attrs['creator'] = f'damask.Result.{creator} v{damask.version}'
 
                 except (OSError,RuntimeError) as err:
-                    print(f'Could not add dataset: {err}.')
+                    logger.error(f'Could not add dataset: {err}.')
 
 
     def _mappings(self):
