### 2025-01-23 #7164 @MarDiehl
@zhubonan: Do you have to run the "accept" again? The link to the final proof points to the draft version that was generated after I asked the bot to do so.
