### 2025-01-23 eeb2279
use "source" and "target" in from_parallel
