### 2025-01-23 b91d51b
too fast, forgot corresponding adjustments in callers of from_parallel
