### 2025-01-25 38ac239
use centralized configuration
