### 2025-01-25 ccb1dbe
Resolve "use NamedTuple in return"
