+++ python/damask/_orientation.py
@@ -1,5 +1,5 @@
 import copy
-from typing import Tuple, Optional, Union, TypeVar, Literal # mypy 1.11, overload
+from typing import Optional, Union, TypeVar, Literal, NamedTuple # mypy 1.11, overload
 
 import numpy as np
 
@@ -11,6 +11,21 @@ from . import tensor
 
 
 MyType = TypeVar('MyType', bound='Orientation')
+
+class DisorientationTuple(NamedTuple):
+    disorientation: 'Orientation'
+    operators: np.ndarray
+
+
+class AverageTuple(NamedTuple):
+    average: 'Orientation'
+    cloud: 'Orientation'
+
+
+class ToSSTTuple(NamedTuple):
+    vector_sst: np.ndarray
+    operator: np.ndarray
+
 
 class Orientation(Rotation,Crystal):
     """
@@ -466,7 +481,7 @@ class Orientation(Rotation,Crystal):
 
     def disorientation(self: MyType,
                        other: MyType,
-                       return_operators: bool = False) -> Union[Tuple[MyType, np.ndarray], MyType]:
+                       return_operators: bool = False) -> Union[MyType, DisorientationTuple]:
         """
         Calculate disorientation between self and given other orientation.
 
@@ -477,7 +492,7 @@ class Orientation(Rotation,Crystal):
             Compatible innermost dimensions will blend.
         return_operators : bool, optional
             Return index pair of symmetrically equivalent orientations
-            that result in disorientation axis falling into FZ.
+            that result in misorientation falling into disorientation FZ.
             Defaults to False.
 
         Returns
@@ -485,7 +500,8 @@ class Orientation(Rotation,Crystal):
         disorientation : Orientation
             Disorientation between self and other.
         operators : numpy.ndarray of int, shape (...,2), conditional
-            Index of symmetrically equivalent orientation that rotated vector to the SST.
+            Index pair of symmetrically equivalent orientations
+            that result in misorientation falling into disorientation FZ.
 
         Notes
         -----
@@ -543,11 +559,11 @@ class Orientation(Rotation,Crystal):
 
         quat = r[ok][sort].reshape((*shp,4))
 
-        return (
-                (self.copy(rotation=quat), (np.vstack(loc[:2]).T)[sort].reshape((*shp,2)))
-                if return_operators else
-                self.copy(rotation=quat)
-               )
+        if return_operators:
+            operators = (np.vstack(loc[:2]).T)[sort].reshape((*shp, 2))
+            return DisorientationTuple(self.copy(rotation=quat), operators)
+        else:
+            return self.copy(rotation=quat)
 
 
     def disorientation_angle(self: MyType,
@@ -637,7 +653,7 @@ class Orientation(Rotation,Crystal):
 
     def average(self: MyType,                                                                       # type: ignore[override]
                 weights: Optional[FloatSequence] = None,
-                return_cloud: bool = False) -> Union[Tuple[MyType, MyType], MyType]:
+                return_cloud: bool = False) -> Union[MyType, AverageTuple]:
         """
         Return orientation average over last dimension.
 
@@ -671,9 +687,10 @@ class Orientation(Rotation,Crystal):
                                                    axis=0),
                                 axis=0))
 
-        return ((self.copy(Rotation(r).average(weights)),self.copy(Rotation(r))) if return_cloud else
-                self.copy(Rotation(r).average(weights))
-               )
+        if return_cloud:
+            return AverageTuple(self.copy(Rotation(r).average(weights)),self.copy(Rotation(r)))
+        else:
+            return self.copy(Rotation(r).average(weights))
 
     # mypy 1.11
     #@overload
@@ -686,7 +703,7 @@ class Orientation(Rotation,Crystal):
[…]
