### 2025-01-26 a68eadf
remember at which increment restart happened
