+++ src/grid/grid_mech_FEM.f90
@@ -435,7 +435,7 @@ subroutine grid_mechanical_FEM_restartWrite()
 
   print'(1x,a)', 'saving solver data required for restart'; flush(IO_STDOUT)
 
-  fileHandle  = HDF5_openFile(getSolverJobName()//'_restart.hdf5','w')
+  fileHandle  = HDF5_openFile(getSolverJobName()//'_restart.hdf5','a')
   groupHandle = HDF5_addGroup(fileHandle,'solver')
   call HDF5_write(reshape(F,[3,3,product(cells(1:2))*cells3]),groupHandle,'F')
   call HDF5_write(reshape(F_lastInc,[3,3,product(cells(1:2))*cells3]),groupHandle,'F_lastInc')
