### 2025-01-28 #7164 @Kevin-Mattheus-Moerman
-> No the order is different and additional authors are listed for the archive. @MarDiehl this one is not a strict requirement in the sense that some folks choose to have true "authors" on the paper, and to add "contributors" to the archive listing. Let me know if you'd like to update the author set on ZENODO to match or if you'd like to leave it as is. I would recommend you change the order of authors to match the paper and to have the paper author set first. Next you could list more as contributors if you like.

-> @MarDiehl Please spell out `USA` as `United States` or `United States of America`.
