### 2025-01-28 901960c
avoid unintroduced abbreviations
