### 2025-01-28 #7164 @MarDiehl
@Kevin-Mattheus-Moerman thanks.

- I've corrected `USA` to `United States of America`.
- The mismatch between Zenodo/GitHub release and the paper is on purpose: DAMASK consists of the main code (in Fortran) and the pre- and post-processing tools written in Python. The release is for the whole package, while the paper is only for the Python tools and only the subset of developer involved in Python coding are listed as authors. This is also a reason for the different ordering.
