+++ paper.bib
@@ -21,7 +21,7 @@
 }
 
 @article{johnstone2020density,
-  title={Density-based clustering of crystal (mis)orientations and the orix Python library},
+  title={Density-based clustering of crystal (mis)orientations and the orix {P}ython library},
   author={Johnstone, Duncan N and Martineau, Ben H and Crout, Phillip and Midgley, Paul A and Eggeman, Alexander S},
   journal={Journal of Applied Crystallography},
   volume={53},
